ui: split app_layout.rs by feature (PR 2.5)
Pure refactor — no behavior change. The 4033-line app_layout.rs is
broken into focused per-feature submodules so subsequent PRs (forge PR
diff mode in particular) have a sane place to add UI without bloating
a single file further.

New modules under src/ui/:
- selector.rs            review-target selector (Local + PR tabs)
- inline_commit_selector.rs  Normal-mode inline commit selector
- file_list.rs           file list panel
- diff_view.rs           shared diff helpers + render_diff_view router
- diff_unified.rs        render_unified_diff + its helpers
- diff_side_by_side.rs   render_side_by_side_diff + its helpers
- text_utils.rs          truncate_str, truncate_or_pad, etc.

app_layout.rs is now ~100 lines and only covers render() entry +
main_content layout routing.

Test count preserved: 459 (selector snapshot tests now live in
src/ui/selector.rs; text + scroll helper tests live with their
functions in text_utils.rs and diff_view.rs).
