+++ src/ui/app_layout.rs
@@ -1,21 +1,15 @@
 use ratatui::{
     Frame,
     layout::{Constraint, Direction, Layout, Rect},
-    style::{Modifier, Style},
-    text::{Line, Span},
-    widgets::{Block, Borders, List, ListItem, Paragraph, Wrap},
+    widgets::Block,
 };
-use unicode_width::UnicodeWidthStr;
 
-use crate::app::{
-    AnnotatedLine, App, DiffViewMode, ExpandDirection, FileTreeItem, FocusedPanel,
-    GAP_EXPAND_BATCH, GapId, InputMode, TargetTab, VisualSelection,
-};
-use crate::forge::selector::{PrTabStatus, PrTabView};
-use crate::model::{LineOrigin, LineRange, LineSide};
-use crate::theme::Theme;
+use crate::app::{App, InputMode};
+use crate::ui::diff_view::render_diff_view;
+use crate::ui::file_list::render_file_list;
+use crate::ui::inline_commit_selector::render_inline_commit_selector;
+use crate::ui::selector::render_commit_select;
 use crate::ui::{comment_panel, help_popup, status_bar, styles};
-use crate::vcs::git::calculate_gap;
 
 pub fn render(frame: &mut Frame, app: &mut App) {
     frame.render_widget(
@@ -75,525 +69,6 @@ pub fn render(frame: &mut Frame, app: &mut App) {
     }
 }
 
-fn render_commit_select(frame: &mut Frame, app: &mut App) {
-    let area = frame.area();
-
-    let chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([
-            Constraint::Length(1), // Header
-            Constraint::Length(1), // Tab strip
-            Constraint::Min(0),    // Active tab body
-            Constraint::Length(1), // Footer hints
-        ])
-        .split(area);
-
-    // Header
-    let header = Paragraph::new(" Select a review target ")
-        .style(styles::header_style(&app.theme))
-        .block(Block::default().style(styles::panel_style(&app.theme)));
-    frame.render_widget(header, chunks[0]);
-
-    render_target_tab_strip(frame, app, chunks[1]);
-
-    match app.target_tab {
-        TargetTab::Local => render_local_target_tab(frame, app, chunks[2]),
-        TargetTab::PullRequests => render_pull_requests_tab(frame, app, chunks[2]),
-    }
-
-    render_target_selector_footer(frame, app, chunks[3]);
-}
-
-fn render_target_tab_strip(frame: &mut Frame, app: &App, area: Rect) {
-    let theme = &app.theme;
-    let active = app.target_tab;
-    let active_style = Style::default()
-        .fg(theme.fg_primary)
-        .add_modifier(Modifier::BOLD | Modifier::REVERSED);
-    let bracket_style = Style::default().fg(theme.fg_secondary);
-    let inactive_label_style = Style::default().fg(theme.fg_primary);
-
-    let mut spans: Vec<Span> = Vec::with_capacity(8);
-    spans.push(Span::raw(" "));
-
-    let local_active = active == TargetTab::Local;
-    spans.push(Span::styled("[", bracket_style));
-    if local_active {
-        spans.push(Span::styled(" Local ", active_style));
-    } else {
-        spans.push(Span::styled(" Local ", inactive_label_style));
-    }
-    spans.push(Span::styled("]", bracket_style));
-
-    spans.push(Span::raw("  "));
-
-    let prs_active = active == TargetTab::PullRequests;
-    spans.push(Span::styled("[", bracket_style));
-    if prs_active {
-        spans.push(Span::styled(" Pull Requests ", active_style));
-    } else {
-        spans.push(Span::styled(" Pull Requests ", inactive_label_style));
-    }
-    spans.push(Span::styled("]", bracket_style));
-
-    let line = Line::from(spans);
-    let strip = Paragraph::new(line).style(styles::panel_style(theme));
-    frame.render_widget(strip, area);
-}
-
-fn render_local_target_tab(frame: &mut Frame, app: &mut App, area: Rect) {
-    let block = Block::default()
-        .title(" Recent Commits ")
-        .borders(Borders::ALL)
-        .style(styles::panel_style(&app.theme))
-        .border_style(styles::border_style(&app.theme, true));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    // Update viewport height for scroll calculations
-    app.commit_list_viewport_height = inner.height as usize;
-    app.commit_list_inner_area = Some(inner);
-    app.pr_list_inner_area = None;
-
-    // Get range info for visual indicators
-    let range = app.commit_selection_range;
-
-    // Determine commits to show
-    let total_commits = app.commit_list.len();
-    let visible_count = app.visible_commit_count.min(total_commits);
-
-    let mut items: Vec<Line> = app
-        .commit_list
-        .iter()
-        .take(visible_count)
-        .enumerate()
-        .map(|(i, commit)| {
-            let is_selected = app.is_commit_selected(i);
-            let is_cursor = i == app.commit_list_cursor;
-
-            // Range boundary indicators
-            let range_marker = match range {
-                Some((start, end)) if i == start && i == end => "─",
-                Some((start, _)) if i == start => "┌",
-                Some((_, end)) if i == end => "└",
-                Some((start, end)) if i > start && i < end => "│",
-                _ => " ",
-            };
-
-            let checkbox = if is_selected { "[x]" } else { "[ ]" };
-            let pointer = if is_cursor { ">" } else { " " };
-
-            let style = if is_cursor {
-                styles::selected_style(&app.theme)
-            } else if is_selected {
-                Style::default().fg(app.theme.fg_secondary)
-            } else {
-                Style::default()
-            };
-
-            let checkbox_style = if is_selected {
-                styles::reviewed_style(&app.theme)
-            } else {
-                styles::pending_style(&app.theme)
-            };
-
-            let range_style = if is_selected {
-                styles::reviewed_style(&app.theme)
-            } else {
-                Style::default().fg(app.theme.fg_secondary)
-            };
-
-            // Format: > ┌ [x] abc1234  Commit message (author, date)
-            let time_str = commit.time.format("%Y-%m-%d").to_string();
-            let mut spans = vec![
-                Span::styled(format!("{pointer} "), style),
-                Span::styled(format!("{range_marker} "), range_style),
-                Span::styled(format!("{checkbox} "), checkbox_style),
-                Span::styled(
-                    format!("{} ", commit.short_id),
-                    styles::hash_style(&app.theme),
-                ),
-            ];
-
-            if commit.id == crate::app::STAGED_SELECTION_ID
-                || commit.id == crate::app::UNSTAGED_SELECTION_ID
-            {
-                spans.push(Span::styled(&commit.summary, style));
-                return Line::from(spans);
-            }
-
-            if let Some(branch_name) = &commit.branch_name {
-                spans.push(Span::styled(
-                    format!("[{}] ", truncate_str(branch_name, 20)),
-                    styles::branch_style(&app.theme),
-                ));
-            }
-
-            spans.push(Span::styled(truncate_str(&commit.summary, 50), style));
-            spans.push(Span::styled(
-                format!(" ({}, {})", commit.author, time_str),
-                Style::default().fg(app.theme.fg_secondary),
-            ));
-
-            Line::from(spans)
-        })
-        .collect();
-
-    // Show an expand row when commits are collapsed
-    if app.can_show_more_commits() {
-        let is_cursor = app.commit_list_cursor == visible_count;
-
-        let style = if is_cursor {
-            styles::selected_style(&app.theme)
-        } else {
-            Style::default().fg(app.theme.fg_secondary)
-        };
-
-        items.push(Line::from(vec![
-            Span::styled(if is_cursor { "> " } else { "  " }, style),
-            Span::styled("       ... show more commits ...", style),
-        ]));
-    }
-
-    // Apply scroll offset and take only visible items
-    let visible_items: Vec<Line> = items
-        .into_iter()
-        .skip(app.commit_list_scroll_offset)
-        .take(inner.height as usize)
-        .collect();
-
-    let list = Paragraph::new(visible_items).style(styles::panel_style(&app.theme));
-    frame.render_widget(list, inner);
-}
-
-fn render_pull_requests_tab(frame: &mut Frame, app: &mut App, area: Rect) {
-    let title = match app.forge_repository.as_ref() {
-        Some(repo) => format!(" Pull Requests ({}) ", repo.display_name()),
-        None => " Pull Requests ".to_string(),
-    };
-    let block = Block::default()
-        .title(title)
-        .borders(Borders::ALL)
-        .style(styles::panel_style(&app.theme))
-        .border_style(styles::border_style(&app.theme, true));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-    app.commit_list_inner_area = None;
-    app.pr_list_inner_area = Some(inner);
-    app.pr_list_viewport_height = inner.height.saturating_sub(1) as usize;
-
-    // Reserve one row at the top for the status / filter banner.
-    let body_chunks = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Length(1), Constraint::Min(0)])
-        .split(inner);
-    let banner_area = body_chunks[0];
-    let list_area = body_chunks[1];
-    app.pr_list_viewport_height = list_area.height as usize;
-
-    let view = app.pr_tab.view();
-    render_pr_banner(frame, app, banner_area, &view);
-    render_pr_list(frame, app, list_area, &view);
-}
-
-fn render_pr_banner(frame: &mut Frame, app: &App, area: Rect, view: &PrTabView<'_>) {
-    let theme = &app.theme;
-
-    if let Some(draft) = app.pr_filter_draft.as_ref() {
-        let prefix = Span::styled(" filter: ", Style::default().fg(theme.fg_secondary));
-        let value = Span::styled(format!("/{draft}"), Style::default().fg(theme.fg_primary));
-        let hint = Span::styled(
-            "   (Enter: apply  Esc: cancel)",
-            Style::default().fg(theme.fg_secondary),
-        );
-        let line = Line::from(vec![prefix, value, hint]);
-        let banner = Paragraph::new(line).style(styles::panel_style(theme));
-        frame.render_widget(banner, area);
-        return;
-    }
-
-    match &view.status {
-        PrTabStatus::Disabled(reason) => {
-            let line = Line::from(Span::styled(
-                format!(" {reason} — switch back with Shift-Tab "),
-                Style::default().fg(theme.fg_secondary),
-            ));
-            frame.render_widget(Paragraph::new(line).style(styles::panel_style(theme)), area);
-        }
-        PrTabStatus::Idle => {
-            let line = Line::from(Span::styled(
-                " Press Tab again or wait — loading…",
-                Style::default().fg(theme.fg_secondary),
-            ));
-            frame.render_widget(Paragraph::new(line).style(styles::panel_style(theme)), area);
-        }
-        PrTabStatus::Loading => {
-            render_pr_progress(frame, app, area, "Fetching pull requests from origin...");
-        }
-        PrTabStatus::LoadingMore => {
-            render_pr_progress(frame, app, area, "Loading more pull requests...");
-        }
-        PrTabStatus::Error(msg) => {
-            let line = Line::from(vec![
-                Span::styled(
-                    " error: ",
-                    Style::default()
-                        .fg(theme.message_error_fg)
-                        .add_modifier(Modifier::BOLD),
-                ),
-                Span::styled((*msg).to_string(), Style::default().fg(theme.fg_primary)),
-            ]);
-            frame.render_widget(Paragraph::new(line).style(styles::panel_style(theme)), area);
-        }
-        PrTabStatus::Ready => {
-            let mut spans: Vec<Span> = vec![Span::styled(
-                format!(" {} loaded", view.rows.len()),
-                Style::default().fg(theme.fg_secondary),
-            )];
-            if !view.filter.is_empty() {
-                spans.push(Span::styled(
-                    format!("   filter: /{}", view.filter),
-                    Style::default().fg(theme.fg_primary),
-                ));
-                spans.push(Span::styled(
-                    "   ('/' to edit, Esc to clear)",
-                    Style::default().fg(theme.fg_secondary),
-                ));
-            } else {
-                spans.push(Span::styled(
-                    "   '/' to filter",
-                    Style::default().fg(theme.fg_secondary),
-                ));
-            }
-            frame.render_widget(
-                Paragraph::new(Line::from(spans)).style(styles::panel_style(theme)),
-                area,
-            );
-        }
-    }
-}
-
-fn render_pr_progress(frame: &mut Frame, app: &App, area: Rect, label: &str) {
-    let theme = &app.theme;
-    let width = area.width.saturating_sub(label.len() as u16 + 5).max(4) as usize;
-    // Indeterminate bar: a windowed block of ~width/3 that travels using
-    // wall-clock seconds, so the bar visibly moves between frames without
-    // requiring per-frame state on App.
-    let block_width = (width / 3).max(2);
-    let travel = (width + block_width).max(1);
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_millis() as usize)
-        .unwrap_or(0);
-    let position = (now / 80) % travel;
-    let mut bar = String::with_capacity(width);
-    for i in 0..width {
-        if position < block_width {
-            if i < position {
-                bar.push(' ');
-            } else if i < position + block_width && i < width {
-                bar.push('#');
-            } else {
-                bar.push(' ');
-            }
-        } else {
-            let start = position.saturating_sub(block_width);
-            if i >= start && i < position && i < width {
-                bar.push('#');
-            } else {
-                bar.push(' ');
-            }
-        }
-    }
-    let line = Line::from(vec![
-        Span::styled(
-            format!(" {label} "),
-            Style::default().fg(theme.fg_secondary),
-        ),
-        Span::styled(
-            format!("[{bar}]"),
-            Style::default().fg(theme.diff_hunk_header),
-        ),
-    ]);
-    frame.render_widget(Paragraph::new(line).style(styles::panel_style(theme)), area);
-}
-
-fn render_pr_list(frame: &mut Frame, app: &App, area: Rect, view: &PrTabView<'_>) {
-    let theme = &app.theme;
-    if area.height == 0 {
-        return;
-    }
-
-    let mut lines: Vec<Line> = Vec::new();
-    for (i, row) in view.rows.iter().enumerate() {
-        let is_cursor = i == view.cursor;
-        let pointer = if is_cursor { "> " } else { "  " };
-        let pointer_style = if is_cursor {
-            styles::selected_style(theme)
-        } else {
-            Style::default().fg(theme.fg_secondary)
-        };
-        let number = format!("#{:<5}", row.summary.number);
-        let title = truncate_str(&row.summary.title, 60);
-        let author = row.summary.author.as_deref().unwrap_or("?");
-        let updated = row
-            .summary
-            .updated_at
-            .map(format_relative_time)
-            .unwrap_or_else(|| "—".to_string());
-        let draft = if row.summary.is_draft { " [draft]" } else { "" };
-
-        let line = Line::from(vec![
-            Span::styled(pointer, pointer_style),
-            Span::styled(number, styles::hash_style(theme)),
-            Span::styled(" ", Style::default()),
-            Span::styled(title, Style::default().fg(theme.fg_primary)),
-            Span::styled(
-                format!("   @{author}"),
-                Style::default().fg(theme.fg_secondary),
-            ),
-            Span::styled(
-                format!("   updated {updated}{draft}"),
-                Style::default().fg(theme.fg_secondary),
-            ),
-        ]);
-        lines.push(line);
-    }
-
-    if view.has_load_more {
-        let load_idx = view.rows.len();
-        let is_cursor = view.cursor == load_idx;
-        let style = if is_cursor {
-            styles::selected_style(theme)
-        } else {
-            Style::default().fg(theme.fg_secondary)
-        };
-        let pointer = if is_cursor { "> " } else { "  " };
-        lines.push(Line::from(vec![
-            Span::styled(pointer, style),
-            Span::styled("... load more pull requests", style),
-        ]));
-    }
-
-    if lines.is_empty() {
-        // Empty state placeholder under the banner.
-        if !matches!(view.status, PrTabStatus::Ready) {
-            return;
-        }
-        let msg = if view.filter.is_empty() {
-            " No open pull requests"
-        } else {
-            " No pull requests match the filter"
-        };
-        lines.push(Line::from(Span::styled(
-            msg,
-            Style::default().fg(theme.fg_secondary),
-        )));
-    }
-
-    let visible: Vec<Line> = lines
-        .into_iter()
-        .skip(view.scroll_offset)
-        .take(area.height as usize)
-        .collect();
-    let paragraph = Paragraph::new(visible).style(styles::panel_style(theme));
-    frame.render_widget(paragraph, area);
-}
-
-fn format_relative_time(time: chrono::DateTime<chrono::Utc>) -> String {
-    let now = chrono::Utc::now();
-    let delta = now.signed_duration_since(time);
-    if delta.num_seconds() < 0 {
-        return "just now".to_string();
-    }
-    let secs = delta.num_seconds();
-    if secs < 60 {
-        return "just now".to_string();
-    }
-    let mins = delta.num_minutes();
-    if mins < 60 {
-        return format!("{mins}m ago");
-    }
-    let hours = delta.num_hours();
-    if hours < 24 {
-        return format!("{hours}h ago");
-    }
-    let days = delta.num_days();
-    if days < 30 {
-        return format!("{days}d ago");
-    }
-    let months = days / 30;
-    if months < 12 {
-        return format!("{months}mo ago");
-    }
-    let years = days / 365;
-    format!("{years}y ago")
-}
-
-fn render_target_selector_footer(frame: &mut Frame, app: &App, area: Rect) {
-    let theme = &app.theme;
-    let mode_span = Span::styled(" SELECT ", styles::mode_style(theme));
-
-    let hints = if app.message.is_some() {
-        String::new()
-    } else if app.pr_filter_editing() {
-        " Enter:apply  Esc:cancel ".to_string()
-    } else {
-        match app.target_tab {
-            TargetTab::Local => {
-                let selected_count = match app.commit_selection_range {
-                    Some((start, end)) => end - start + 1,
-                    None => 0,
-                };
-                let selection_info = if selected_count > 0 {
-                    format!(" ({selected_count} selected)")
-                } else {
-                    String::new()
-                };
-                format!(
-                    " Tab:tabs  j/k:navigate  Space:select range  Enter:confirm  q:quit{selection_info}"
-                )
-            }
-            TargetTab::PullRequests => {
-                " Tab:tabs  j/k:navigate  Enter:open  /:filter  Esc/q:back ".to_string()
-            }
-        }
-    };
-    let hints_span = Span::styled(hints, Style::default().fg(theme.fg_secondary));
-
-    let left_spans = vec![mode_span, hints_span];
-
-    let (message_span, message_width) = status_bar::build_message_span(app.message.as_ref(), theme);
-    let spans = status_bar::build_right_aligned_spans(
-        left_spans,
-        message_span,
-        message_width,
-        area.width as usize,
-    );
-
-    let footer = Paragraph::new(Line::from(spans))
-        .style(styles::status_bar_style(theme))
-        .block(Block::default());
-    frame.render_widget(footer, area);
-}
-
-fn truncate_str(s: &str, max_len: usize) -> String {
-    if s.len() <= max_len {
-        s.to_string()
-    } else {
-        let truncate_at = max_len.saturating_sub(3);
-        let end = s
-            .char_indices()
-            .map(|(i, _)| i)
-            .take_while(|&i| i <= truncate_at)
-            .last()
-            .unwrap_or(0);
-        format!("{}...", &s[..end])
-    }
-}
-
 fn render_main_content(frame: &mut Frame, app: &mut App, area: Rect) {
     let content_area = if app.has_inline_commit_selector() {
         let selector_height = (app.review_commits.len() as u16 + 2).min(8); // N items + 2 borders, capped
@@ -627,3407 +102,5 @@ fn render_main_content(frame: &mut Frame, app: &mut App, area: Rect) {
         app.diff_area = Some(content_area);
 
         render_diff_view(frame, app, content_area);
-    }
-}
-
-fn render_inline_commit_selector(frame: &mut Frame, app: &mut App, area: Rect) {
-    let focused = app.focused_panel == FocusedPanel::CommitSelector;
-    let block = Block::default()
-        .title(" Commits ")
-        .borders(Borders::ALL)
-        .border_style(styles::border_style(&app.theme, focused));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    // Update viewport height for scroll
-    app.commit_list_viewport_height = inner.height as usize;
-    app.commit_list_inner_area = Some(inner);
-
-    {
-        let range = app.commit_selection_range;
-        let total_commits = app.review_commits.len();
-
-        let items: Vec<Line> = app
-            .review_commits
-            .iter()
-            .take(total_commits)
-            .enumerate()
-            .map(|(i, commit)| {
-                let is_selected = app.is_commit_selected(i);
-                let is_cursor = i == app.commit_list_cursor;
-
-                // Range boundary indicators
-                let range_marker = match range {
-                    Some((start, end)) if i == start && i == end => "\u{2500}",
-                    Some((start, _)) if i == start => "\u{250c}",
-                    Some((_, end)) if i == end => "\u{2514}",
-                    Some((start, end)) if i > start && i < end => "\u{2502}",
-                    _ => " ",
-                };
-
-                let checkbox = if is_selected { "[x]" } else { "[ ]" };
-
-                let style = if is_cursor {
-                    styles::selected_style(&app.theme)
-                } else if is_selected {
-                    Style::default().fg(app.theme.fg_secondary)
-                } else {
-                    Style::default()
-                };
-
-                let checkbox_style = if is_selected {
-                    styles::reviewed_style(&app.theme)
-                } else {
-                    styles::pending_style(&app.theme)
-                };
-
-                let range_style = if is_selected {
-                    styles::reviewed_style(&app.theme)
-                } else {
-                    Style::default().fg(app.theme.fg_secondary)
-                };
-
-                let pointer = if is_cursor { "> " } else { "  " };
-
-                let time_str = commit.time.format("%Y-%m-%d").to_string();
-                let mut spans = vec![
-                    Span::styled(pointer.to_string(), style),
-                    Span::styled(format!("{} ", range_marker), range_style),
-                    Span::styled(format!("{} ", checkbox), checkbox_style),
-                    Span::styled(
-                        format!("{} ", commit.short_id),
-                        styles::hash_style(&app.theme),
-                    ),
-                ];
-
-                if let Some(branch_name) = &commit.branch_name {
-                    spans.push(Span::styled(
-                        format!("[{}] ", truncate_str(branch_name, 20)),
-                        styles::branch_style(&app.theme),
-                    ));
-                }
-
-                spans.push(Span::styled(truncate_str(&commit.summary, 50), style));
-                spans.push(Span::styled(
-                    format!(" ({}, {})", commit.author, time_str),
-                    Style::default().fg(app.theme.fg_secondary),
-                ));
-
-                Line::from(spans)
-            })
-            .collect();
-
-        let visible_items: Vec<Line> = items
-            .into_iter()
-            .skip(app.commit_list_scroll_offset)
-            .take(inner.height as usize)
-            .collect();
-
-        let paragraph = Paragraph::new(visible_items);
-        frame.render_widget(paragraph, inner);
-    }
-}
-
-fn render_file_list(frame: &mut Frame, app: &mut App, area: Rect) {
-    use ratatui::style::Modifier;
-    use std::path::Path;
-
-    let focused = app.focused_panel == FocusedPanel::FileList;
-
-    let block = Block::default()
-        .title(" Files ")
-        .borders(Borders::ALL)
-        .style(styles::panel_style(&app.theme))
-        .border_style(styles::border_style(&app.theme, focused));
-
-    let inner = block.inner(area);
-    app.file_list_inner_area = Some(inner);
-    let visible_items = app.build_visible_items();
-
-    let max_content_width = visible_items
-        .iter()
-        .map(|item| match item {
-            FileTreeItem::Directory { path, depth, .. } => {
-                let dir_name = Path::new(path)
-                    .file_name()
-                    .and_then(|n| n.to_str())
-                    .unwrap_or(path);
-                depth * 2 + 2 + dir_name.width() + 1
-            }
-            FileTreeItem::File { file_idx, depth } => {
-                let file = &app.diff_files[*file_idx];
-                let filename = file
-                    .display_path()
-                    .file_name()
-                    .and_then(|n| n.to_str())
-                    .unwrap_or("?");
-                depth * 2 + 3 + 3 + filename.width()
-            }
-        })
-        .max()
-        .unwrap_or(0);
-
-    app.file_list_state.viewport_width = inner.width as usize;
-    app.file_list_state.viewport_height = inner.height as usize;
-    app.file_list_state.max_content_width = max_content_width;
-
-    let max_scroll_x = max_content_width.saturating_sub(inner.width as usize);
-    if app.file_list_state.scroll_x > max_scroll_x {
-        app.file_list_state.scroll_x = max_scroll_x;
-    }
-    let scroll_x = app.file_list_state.scroll_x;
-
-    // When diff panel is focused, sync file list selection to current file
-    // But preserve the current offset to not interfere with manual scrolling
-    if app.focused_panel == FocusedPanel::Diff {
-        let current_file_idx = app.diff_state.current_file_idx;
-        for (tree_idx, item) in visible_items.iter().enumerate() {
-            if let FileTreeItem::File { file_idx, .. } = item
-                && *file_idx == current_file_idx
-            {
-                if app.file_list_state.selected() != tree_idx {
-                    // Save current offset before changing selection
-                    let current_offset = app.file_list_state.list_state.offset();
-                    app.file_list_state.select(tree_idx);
-                    // Restore offset to prevent auto-scrolling
-                    *app.file_list_state.list_state.offset_mut() = current_offset;
-                }
-                break;
-            }
-        }
-    }
-
-    let selected_idx = app.file_list_state.selected();
-
-    let items: Vec<ListItem> = visible_items
-        .iter()
-        .enumerate()
-        .map(|(i, item)| {
-            let is_selected = i == selected_idx;
-
-            match item {
-                FileTreeItem::Directory {
-                    path,
-                    depth,
-                    expanded,
-                } => {
-                    let indent = "  ".repeat(*depth);
-                    let icon = if *expanded { "▾" } else { "▸" };
-                    let dir_name = Path::new(path)
-                        .file_name()
-                        .and_then(|n| n.to_str())
-                        .unwrap_or(path);
-
-                    let style = if is_selected {
-                        styles::selected_style(&app.theme).add_modifier(Modifier::UNDERLINED)
-                    } else {
-                        Style::default()
-                    };
-
-                    let line = Line::from(vec![
-                        Span::styled(indent, Style::default()),
-                        Span::styled(format!("{icon} "), styles::dir_icon_style(&app.theme)),
-                        Span::styled(format!("{dir_name}/"), style),
-                    ]);
-
-                    ListItem::new(apply_horizontal_scroll(line, scroll_x))
-                }
-                FileTreeItem::File { file_idx, depth } => {
-                    let file = &app.diff_files[*file_idx];
-                    let path = file.display_path();
-                    let is_reviewed = app.session.is_file_reviewed(path);
-                    let review_mark = if is_reviewed { "✓" } else { " " };
-
-                    let style = if is_selected {
-                        styles::selected_style(&app.theme).add_modifier(Modifier::UNDERLINED)
-                    } else {
-                        Style::default()
-                    };
-
-                    let line = if file.is_commit_message {
-                        Line::from(vec![
-                            Span::styled(
-                                format!("[{review_mark}]"),
-                                if is_reviewed {
-                                    styles::reviewed_style(&app.theme)
-                                } else {
-                                    styles::pending_style(&app.theme)
-                                },
-                            ),
-                            Span::styled("   Commit Message".to_string(), style),
-                        ])
-                    } else {
-                        let filename = path.file_name().and_then(|n| n.to_str()).unwrap_or("?");
-                        let status = file.status.as_char();
-                        let indent = "  ".repeat(*depth);
-                        Line::from(vec![
-                            Span::styled(indent, Style::default()),
-                            Span::styled(
-                                format!("[{review_mark}]"),
-                                if is_reviewed {
-                                    styles::reviewed_style(&app.theme)
-                                } else {
-                                    styles::pending_style(&app.theme)
-                                },
-                            ),
-                            Span::styled(
-                                format!(" {status} "),
-                                styles::file_status_style(&app.theme, status),
-                            ),
-                            Span::styled(filename.to_string(), style),
-                        ])
-                    };
-
-                    ListItem::new(apply_horizontal_scroll(line, scroll_x))
-                }
-            }
-        })
-        .collect();
-
-    let list = List::new(items)
-        .style(styles::panel_style(&app.theme))
-        .block(block);
-
-    frame.render_stateful_widget(list, area, &mut app.file_list_state.list_state);
-}
-
-fn render_diff_view(frame: &mut Frame, app: &mut App, area: Rect) {
-    match app.diff_view_mode {
-        DiffViewMode::Unified => render_unified_diff(frame, app, area),
-        DiffViewMode::SideBySide => render_side_by_side_diff(frame, app, area),
-    }
-}
-
-/// Build a right-aligned title showing diff stats for the current scope.
-/// In overview: total stats across all files. In a file: that file's stats.
-fn diff_stat_title(app: &App) -> Line<'static> {
-    let (additions, deletions) = if app.is_cursor_in_overview() || app.current_file_path().is_none()
-    {
-        let (_, a, d) = app.diff_stat();
-        (a, d)
-    } else {
-        app.diff_files[app.diff_state.current_file_idx].stat()
-    };
-
-    let theme = &app.theme;
-    Line::from(vec![
-        Span::styled(
-            format!(" +{additions}"),
-            Style::default().fg(theme.diff_add),
-        ),
-        Span::raw(" "),
-        Span::styled(
-            format!("-{deletions} "),
-            Style::default().fg(theme.diff_del),
-        ),
-    ])
-}
-
-fn render_unified_diff(frame: &mut Frame, app: &mut App, area: Rect) {
-    let focused = app.focused_panel == FocusedPanel::Diff;
-
-    let title = if app.is_cursor_in_overview() || app.current_file_path().is_none() {
-        " Diff (Unified) \u{2014} Overview ".to_string()
-    } else {
-        format!(
-            " Diff (Unified) \u{2014} {} ",
-            app.current_file_path().unwrap().display()
-        )
-    };
-
-    let block = Block::default()
-        .title(title)
-        .title_top(diff_stat_title(app).right_aligned())
-        .borders(Borders::ALL)
-        .style(styles::panel_style(&app.theme))
-        .border_style(styles::border_style(&app.theme, focused));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    // Update viewport height for scroll calculations
-    app.diff_state.viewport_height = inner.height as usize;
-    app.diff_inner_area = Some(inner);
-
-    // Reset comment input annotation offset (will be set if a comment input box is rendered)
-    app.comment_input_annotation_offset = None;
-
-    // Build all diff lines for infinite scroll
-    // Track line index to mark the current line (cursor position)
-    let mut lines: Vec<Line> = Vec::new();
-    let mut line_idx: usize = 0;
-    let current_line_idx = app.diff_state.cursor_line;
-
-    // Track cursor position for IME when in Comment mode
-    // Store the logical line index and column where the cursor should be
-    let mut comment_cursor_logical_line: Option<usize> = None;
-    let mut comment_cursor_column: u16 = 0;
-    // Track the full extent of the comment input box so we can auto-scroll
-    // the viewport to keep it visible while the user types.
-    let mut comment_input_box_range: Option<(usize, usize)> = None;
-
-    let is_review_comment_mode =
-        app.input_mode == InputMode::Comment && app.comment_is_review_level;
-
-    let general_indicator = cursor_indicator_spaced(line_idx, current_line_idx);
-    lines.push(Line::from(vec![
-        Span::styled(
-            general_indicator,
-            styles::current_line_indicator_style(&app.theme),
-        ),
-        Span::styled(
-            "═══ Review Comments ",
-            styles::file_header_style(&app.theme),
-        ),
-        Span::styled("═".repeat(40), styles::file_header_style(&app.theme)),
-    ]));
-    line_idx += 1;
-
-    for comment in &app.session.review_comments {
-        let is_being_edited =
-            app.editing_comment_id.as_ref() == Some(&comment.id) && is_review_comment_mode;
-
-        if is_being_edited {
-            let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-                &app.theme,
-                comment_type_presentation(app, &app.comment_type),
-                &app.comment_buffer,
-                app.comment_cursor,
-                None,
-                true,
-                app.supports_keyboard_enhancement,
-            );
-            comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-            comment_cursor_column = 1 + cursor_info.column;
-            comment_input_box_range =
-                Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-            let annotations_replaced = 2 + comment.content.split('\n').count();
-            app.comment_input_annotation_offset =
-                Some((line_idx, input_lines.len(), annotations_replaced));
-
-            for mut input_line in input_lines {
-                let indicator = cursor_indicator(line_idx, current_line_idx);
-                input_line.spans.insert(
-                    0,
-                    Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                );
-                lines.push(input_line);
-                line_idx += 1;
-            }
-        } else {
-            let comment_lines = comment_panel::format_comment_lines(
-                &app.theme,
-                comment_type_presentation(app, &comment.comment_type),
-                &comment.content,
-                None,
-            );
-            for mut comment_line in comment_lines {
-                let indicator = cursor_indicator(line_idx, current_line_idx);
-                comment_line.spans.insert(
-                    0,
-                    Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                );
-                lines.push(comment_line);
-                line_idx += 1;
-            }
-        }
-    }
-
-    if is_review_comment_mode && app.editing_comment_id.is_none() {
-        let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-            &app.theme,
-            comment_type_presentation(app, &app.comment_type),
-            &app.comment_buffer,
-            app.comment_cursor,
-            None,
-            false,
-            app.supports_keyboard_enhancement,
-        );
-        comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-        comment_cursor_column = 1 + cursor_info.column;
-        comment_input_box_range = Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-        app.comment_input_annotation_offset = Some((line_idx, input_lines.len(), 0));
-
-        for mut input_line in input_lines {
-            let indicator = cursor_indicator(line_idx, current_line_idx);
-            input_line.spans.insert(
-                0,
-                Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-            );
-            lines.push(input_line);
-            line_idx += 1;
-        }
-    }
-
-    for (file_idx, file) in app.diff_files.iter().enumerate() {
-        let path = file.display_path();
-        let status = file.status.as_char();
-        let is_reviewed = app.session.is_file_reviewed(path);
-
-        // File header
-        let indicator = cursor_indicator_spaced(line_idx, current_line_idx);
-
-        // Add checkmark if reviewed (using same character as file list)
-        let review_mark = if is_reviewed { "✓ " } else { "" };
-
-        let header_text = if file.is_commit_message {
-            format!("═══ {}Commit Message ", review_mark)
-        } else {
-            format!("═══ {}{} [{}] ", review_mark, path.display(), status)
-        };
-        lines.push(Line::from(vec![
-            Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-            Span::styled(header_text, styles::file_header_style(&app.theme)),
-            Span::styled("═".repeat(40), styles::file_header_style(&app.theme)),
-        ]));
-        line_idx += 1;
-
-        // If file is reviewed, skip rendering the body (fold it away)
-        if is_reviewed {
-            continue;
-        }
-
-        // Check if we're editing/adding a file-level comment for this file
-        let is_file_comment_mode = app.input_mode == InputMode::Comment
-            && app.comment_is_file_level
-            && file_idx == app.diff_state.current_file_idx;
-
-        // Show file-level comments right after the header
-        if let Some(review) = app.session.files.get(path) {
-            for comment in &review.file_comments {
-                // Skip rendering this comment if it's being edited
-                let is_being_edited =
-                    app.editing_comment_id.as_ref() == Some(&comment.id) && is_file_comment_mode;
-
-                if is_being_edited {
-                    // Render the inline input instead
-                    let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-                        &app.theme,
-                        comment_type_presentation(app, &app.comment_type),
-                        &app.comment_buffer,
-                        app.comment_cursor,
-                        None,
-                        true,
-                        app.supports_keyboard_enhancement,
-                    );
-                    // Track cursor position: logical line = current line_idx + cursor offset within input
-                    comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-                    // Column = indicator (1) + cursor_info.column
-                    comment_cursor_column = 1 + cursor_info.column;
-                    comment_input_box_range =
-                        Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-                    let annotations_replaced = 2 + comment.content.split('\n').count();
-                    app.comment_input_annotation_offset =
-                        Some((line_idx, input_lines.len(), annotations_replaced));
-
-                    for mut input_line in input_lines {
-                        let indicator = cursor_indicator(line_idx, current_line_idx);
-                        input_line.spans.insert(
-                            0,
-                            Span::styled(
-                                indicator,
-                                styles::current_line_indicator_style(&app.theme),
-                            ),
-                        );
-                        lines.push(input_line);
-                        line_idx += 1;
-                    }
-                } else {
-                    let comment_lines = comment_panel::format_comment_lines(
-                        &app.theme,
-                        comment_type_presentation(app, &comment.comment_type),
-                        &comment.content,
-                        None,
-                    );
-                    for mut comment_line in comment_lines {
-                        let indicator = cursor_indicator(line_idx, current_line_idx);
-                        comment_line.spans.insert(
-                            0,
-                            Span::styled(
-                                indicator,
-                                styles::current_line_indicator_style(&app.theme),
-                            ),
-                        );
-                        lines.push(comment_line);
-                        line_idx += 1;
-                    }
-                }
-            }
-        }
-
-        // Render inline input for new file-level comment
-        if is_file_comment_mode && app.editing_comment_id.is_none() {
-            let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-                &app.theme,
-                comment_type_presentation(app, &app.comment_type),
-                &app.comment_buffer,
-                app.comment_cursor,
-                None,
-                false,
-                app.supports_keyboard_enhancement,
-            );
-            // Track cursor position
-            comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-            comment_cursor_column = 1 + cursor_info.column;
-            comment_input_box_range =
-                Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-            app.comment_input_annotation_offset = Some((line_idx, input_lines.len(), 0));
-
-            for mut input_line in input_lines {
-                let indicator = cursor_indicator(line_idx, current_line_idx);
-                input_line.spans.insert(
-                    0,
-                    Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                );
-                lines.push(input_line);
-                line_idx += 1;
-            }
-        }
-
-        if file.is_too_large {
-            let indicator = cursor_indicator_spaced(line_idx, current_line_idx);
-            lines.push(Line::from(vec![
-                Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                Span::styled("(file too large to display)", styles::dim_style(&app.theme)),
-            ]));
-            line_idx += 1;
-        } else if file.is_binary {
-            let indicator = cursor_indicator_spaced(line_idx, current_line_idx);
-            lines.push(Line::from(vec![
-                Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                Span::styled("(binary file)", styles::dim_style(&app.theme)),
-            ]));
-            line_idx += 1;
-        } else if file.hunks.is_empty() {
-            let indicator = cursor_indicator_spaced(line_idx, current_line_idx);
-            lines.push(Line::from(vec![
-                Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                Span::styled("(no changes)", styles::dim_style(&app.theme)),
-            ]));
-            line_idx += 1;
-        } else {
-            // Get line comments for this file
-            let line_comments = app
-                .session
-                .files
-                .get(path)
-                .map(|r| &r.line_comments)
-                .cloned()
-                .unwrap_or_default();
-
-            for (hunk_idx, hunk) in file.hunks.iter().enumerate() {
-                // Calculate and render gap before this hunk
-                let prev_hunk = if hunk_idx > 0 {
-                    file.hunks.get(hunk_idx - 1)
-                } else {
-                    None
-                };
-                let gap = calculate_gap(
-                    prev_hunk.map(|h| (&h.new_start, &h.new_count)),
-                    hunk.new_start,
-                );
-
-                let gap_id = GapId { file_idx, hunk_idx };
-
-                if gap > 0 {
-                    let top_lines = app.expanded_top.get(&gap_id);
-                    let bot_lines = app.expanded_bottom.get(&gap_id);
-                    let top_len = top_lines.map_or(0, |v| v.len());
-                    let bot_len = bot_lines.map_or(0, |v| v.len());
-                    let remaining = (gap as usize).saturating_sub(top_len + bot_len);
-                    let is_top_of_file = hunk_idx == 0;
-
-                    // Render top expanded lines
-                    if let Some(top) = top_lines {
-                        for expanded_line in top {
-                            render_expanded_context_line(
-                                &mut lines,
-                                &mut line_idx,
-                                current_line_idx,
-                                expanded_line,
-                                &app.theme,
-                            );
-                        }
-                    }
-
-                    // Render expanders / hidden lines
-                    if remaining > 0 {
-                        if is_top_of_file {
-                            if remaining > GAP_EXPAND_BATCH {
-                                render_hidden_lines(
-                                    &mut lines,
-                                    &mut line_idx,
-                                    current_line_idx,
-                                    remaining,
-                                    &app.theme,
-                                );
-                            }
-                            render_expander_line(
-                                &mut lines,
-                                &mut line_idx,
-                                current_line_idx,
-                                ExpandDirection::Up,
-                                remaining,
-                                &app.theme,
-                            );
-                        } else if remaining >= GAP_EXPAND_BATCH {
-                            render_expander_line(
-                                &mut lines,
-                                &mut line_idx,
-                                current_line_idx,
-                                ExpandDirection::Down,
-                                remaining,
-                                &app.theme,
-                            );
-                            render_hidden_lines(
-                                &mut lines,
-                                &mut line_idx,
-                                current_line_idx,
-                                remaining,
-                                &app.theme,
-                            );
-                            render_expander_line(
-                                &mut lines,
-                                &mut line_idx,
-                                current_line_idx,
-                                ExpandDirection::Up,
-                                remaining,
-                                &app.theme,
-                            );
-                        } else {
-                            render_expander_line(
-                                &mut lines,
-                                &mut line_idx,
-                                current_line_idx,
-                                ExpandDirection::Both,
-                                remaining,
-                                &app.theme,
-                            );
-                        }
-                    }
-
-                    // Render bottom expanded lines
-                    if let Some(bot) = bot_lines {
-                        for expanded_line in bot {
-                            render_expanded_context_line(
-                                &mut lines,
-                                &mut line_idx,
-                                current_line_idx,
-                                expanded_line,
-                                &app.theme,
-                            );
-                        }
-                    }
-                }
-
-                // Hunk header
-                let indicator = cursor_indicator_spaced(line_idx, current_line_idx);
-                lines.push(Line::from(vec![
-                    Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                    Span::styled(
-                        hunk.header.to_string(),
-                        styles::diff_hunk_header_style(&app.theme),
-                    ),
-                ]));
-                line_idx += 1;
-
-                // Diff lines
-                for diff_line in &hunk.lines {
-                    let (prefix, base_style) = match diff_line.origin {
-                        LineOrigin::Addition => ("+", styles::diff_add_style(&app.theme)),
-                        LineOrigin::Deletion => ("-", styles::diff_del_style(&app.theme)),
-                        LineOrigin::Context => (" ", styles::diff_context_style(&app.theme)),
-                    };
-
-                    let style = base_style;
-
-                    let line_num_str = match diff_line.origin {
-                        LineOrigin::Addition => diff_line
-                            .new_lineno
-                            .map(|n| format!("{n:>4} "))
-                            .unwrap_or_else(|| "     ".to_string()),
-                        LineOrigin::Deletion => diff_line
-                            .old_lineno
-                            .map(|n| format!("{n:>4} "))
-                            .unwrap_or_else(|| "     ".to_string()),
-                        _ => diff_line
-                            .new_lineno
-                            .or(diff_line.old_lineno)
-                            .map(|n| format!("{n:>4} "))
-                            .unwrap_or_else(|| "     ".to_string()),
-                    };
-
-                    let indicator = cursor_indicator(line_idx, current_line_idx);
-
-                    let line_num_style = styles::dim_style(&app.theme);
-
-                    let mut line_spans = vec![
-                        Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                        Span::styled(line_num_str, line_num_style),
-                        Span::styled(format!("{prefix} "), style),
-                    ];
-
-                    if let Some(ref highlighted) = diff_line.highlighted_spans {
-                        for (span_style, span_text) in highlighted {
-                            line_spans.push(Span::styled(span_text.clone(), *span_style));
-                        }
-                    } else {
-                        line_spans.push(Span::styled(diff_line.content.clone(), style));
-                    }
-
-                    // Mark add/del lines with their effective EOL style so we can paint full
-                    // row backgrounds later (including wrapped visual rows).
-                    if matches!(
-                        diff_line.origin,
-                        LineOrigin::Addition | LineOrigin::Deletion
-                    ) {
-                        let eol_style = match diff_line.highlighted_spans.as_ref() {
-                            // For syntax-highlighted lines (including empty highlighted lines),
-                            // use syntax diff background so row fill matches code spans.
-                            Some(_) => {
-                                let syntax_bg = match diff_line.origin {
-                                    LineOrigin::Addition => app.theme.syntax_add_bg,
-                                    LineOrigin::Deletion => app.theme.syntax_del_bg,
-                                    LineOrigin::Context => app.theme.panel_bg,
-                                };
-                                let base = line_spans.last().map(|s| s.style).unwrap_or(style);
-                                base.bg(syntax_bg)
-                            }
-                            // Non-highlighted lines keep classic diff background.
-                            None => line_spans.last().map(|s| s.style).unwrap_or(style),
-                        };
-                        // Zero-width marker span carrying the background style.
-                        line_spans.push(Span::styled(String::new(), eol_style));
-                    }
-
-                    lines.push(Line::from(line_spans));
-                    line_idx += 1;
-
-                    // Show line comments for both old side (deleted lines) and new side (added/context)
-                    // Old side comments (for deleted lines)
-                    if let Some(old_ln) = diff_line.old_lineno {
-                        // Check if we're adding/editing a comment on this line (old side)
-                        let is_line_comment_mode = app.input_mode == InputMode::Comment
-                            && !app.comment_is_file_level
-                            && file_idx == app.diff_state.current_file_idx
-                            && app.comment_line == Some((old_ln, LineSide::Old));
-
-                        if let Some(comments) = line_comments.get(&old_ln) {
-                            for comment in comments {
-                                if comment.side == Some(LineSide::Old) {
-                                    // Skip if this comment is being edited
-                                    let is_being_edited = is_line_comment_mode
-                                        && app.editing_comment_id.as_ref() == Some(&comment.id);
-
-                                    if is_being_edited {
-                                        let line_range = app
-                                            .comment_line_range
-                                            .map(|(r, _)| r)
-                                            .or_else(|| Some(LineRange::single(old_ln)));
-                                        let (input_lines, cursor_info) =
-                                            comment_panel::format_comment_input_lines(
-                                                &app.theme,
-                                                comment_type_presentation(app, &app.comment_type),
-                                                &app.comment_buffer,
-                                                app.comment_cursor,
-                                                line_range,
-                                                true,
-                                                app.supports_keyboard_enhancement,
-                                            );
-                                        comment_cursor_logical_line =
-                                            Some(line_idx + cursor_info.line_offset);
-                                        comment_cursor_column = 1 + cursor_info.column;
-                                        comment_input_box_range = Some((
-                                            line_idx,
-                                            line_idx + input_lines.len().saturating_sub(1),
-                                        ));
-                                        let annotations_replaced =
-                                            2 + comment.content.split('\n').count();
-                                        app.comment_input_annotation_offset = Some((
-                                            line_idx,
-                                            input_lines.len(),
-                                            annotations_replaced,
-                                        ));
-
-                                        for mut input_line in input_lines {
-                                            let indicator =
-                                                cursor_indicator(line_idx, current_line_idx);
-                                            input_line.spans.insert(
-                                                0,
-                                                Span::styled(
-                                                    indicator,
-                                                    styles::current_line_indicator_style(
-                                                        &app.theme,
-                                                    ),
-                                                ),
-                                            );
-                                            lines.push(input_line);
-                                            line_idx += 1;
-                                        }
-                                    } else {
-                                        let line_range = comment
-                                            .line_range
-                                            .or_else(|| Some(LineRange::single(old_ln)));
-                                        let comment_lines = comment_panel::format_comment_lines(
-                                            &app.theme,
-                                            comment_type_presentation(app, &comment.comment_type),
-                                            &comment.content,
-                                            line_range,
-                                        );
-                                        for mut comment_line in comment_lines {
-                                            let is_current = line_idx == current_line_idx;
-                                            let indicator = if is_current { "▶" } else { " " };
-                                            comment_line.spans.insert(
-                                                0,
-                                                Span::styled(
-                                                    indicator,
-                                                    styles::current_line_indicator_style(
-                                                        &app.theme,
-                                                    ),
-                                                ),
-                                            );
-                                            lines.push(comment_line);
-                                            line_idx += 1;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-
-                        // Render inline input for new line comment (old side)
-                        if is_line_comment_mode && app.editing_comment_id.is_none() {
-                            let line_range = app
-                                .comment_line_range
-                                .map(|(r, _)| r)
-                                .or_else(|| Some(LineRange::single(old_ln)));
-                            let (input_lines, cursor_info) =
-                                comment_panel::format_comment_input_lines(
-                                    &app.theme,
-                                    comment_type_presentation(app, &app.comment_type),
-                                    &app.comment_buffer,
-                                    app.comment_cursor,
-                                    line_range,
-                                    false,
-                                    app.supports_keyboard_enhancement,
-                                );
-                            comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-                            comment_cursor_column = 1 + cursor_info.column;
-                            comment_input_box_range =
-                                Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-                            app.comment_input_annotation_offset =
-                                Some((line_idx, input_lines.len(), 0));
-
-                            for mut input_line in input_lines {
-                                let indicator = cursor_indicator(line_idx, current_line_idx);
-                                input_line.spans.insert(
-                                    0,
-                                    Span::styled(
-                                        indicator,
-                                        styles::current_line_indicator_style(&app.theme),
-                                    ),
-                                );
-                                lines.push(input_line);
-                                line_idx += 1;
-                            }
-                        }
-                    }
-
-                    // New side comments (for added/context lines)
-                    if let Some(new_ln) = diff_line.new_lineno {
-                        // Check if we're adding/editing a comment on this line (new side)
-                        let is_line_comment_mode = app.input_mode == InputMode::Comment
-                            && !app.comment_is_file_level
-                            && file_idx == app.diff_state.current_file_idx
-                            && app.comment_line == Some((new_ln, LineSide::New));
-
-                        if let Some(comments) = line_comments.get(&new_ln) {
-                            for comment in comments {
-                                if comment.side != Some(LineSide::Old) {
-                                    // Skip if this comment is being edited
-                                    let is_being_edited = is_line_comment_mode
-                                        && app.editing_comment_id.as_ref() == Some(&comment.id);
-
-                                    if is_being_edited {
-                                        let line_range = app
-                                            .comment_line_range
-                                            .map(|(r, _)| r)
-                                            .or_else(|| Some(LineRange::single(new_ln)));
-                                        let (input_lines, cursor_info) =
-                                            comment_panel::format_comment_input_lines(
-                                                &app.theme,
-                                                comment_type_presentation(app, &app.comment_type),
-                                                &app.comment_buffer,
-                                                app.comment_cursor,
-                                                line_range,
-                                                true,
-                                                app.supports_keyboard_enhancement,
-                                            );
-                                        comment_cursor_logical_line =
-                                            Some(line_idx + cursor_info.line_offset);
-                                        comment_cursor_column = 1 + cursor_info.column;
-                                        comment_input_box_range = Some((
-                                            line_idx,
-                                            line_idx + input_lines.len().saturating_sub(1),
-                                        ));
-                                        let annotations_replaced =
-                                            2 + comment.content.split('\n').count();
-                                        app.comment_input_annotation_offset = Some((
-                                            line_idx,
-                                            input_lines.len(),
-                                            annotations_replaced,
-                                        ));
-
-                                        for mut input_line in input_lines {
-                                            let indicator =
-                                                cursor_indicator(line_idx, current_line_idx);
-                                            input_line.spans.insert(
-                                                0,
-                                                Span::styled(
-                                                    indicator,
-                                                    styles::current_line_indicator_style(
-                                                        &app.theme,
-                                                    ),
-                                                ),
-                                            );
-                                            lines.push(input_line);
-                                            line_idx += 1;
-                                        }
-                                    } else {
-                                        let line_range = comment
-                                            .line_range
-                                            .or_else(|| Some(LineRange::single(new_ln)));
-                                        let comment_lines = comment_panel::format_comment_lines(
-                                            &app.theme,
-                                            comment_type_presentation(app, &comment.comment_type),
-                                            &comment.content,
-                                            line_range,
-                                        );
-                                        for mut comment_line in comment_lines {
-                                            let indicator =
-                                                cursor_indicator(line_idx, current_line_idx);
-                                            comment_line.spans.insert(
-                                                0,
-                                                Span::styled(
-                                                    indicator,
-                                                    styles::current_line_indicator_style(
-                                                        &app.theme,
-                                                    ),
-                                                ),
-                                            );
-                                            lines.push(comment_line);
-                                            line_idx += 1;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-
-                        // Render inline input for new line comment (new side)
-                        if is_line_comment_mode && app.editing_comment_id.is_none() {
-                            let line_range = app
-                                .comment_line_range
-                                .map(|(r, _)| r)
-                                .or_else(|| Some(LineRange::single(new_ln)));
-                            let (input_lines, cursor_info) =
-                                comment_panel::format_comment_input_lines(
-                                    &app.theme,
-                                    comment_type_presentation(app, &app.comment_type),
-                                    &app.comment_buffer,
-                                    app.comment_cursor,
-                                    line_range,
-                                    false,
-                                    app.supports_keyboard_enhancement,
-                                );
-                            comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-                            comment_cursor_column = 1 + cursor_info.column;
-                            comment_input_box_range =
-                                Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-                            app.comment_input_annotation_offset =
-                                Some((line_idx, input_lines.len(), 0));
-
-                            for mut input_line in input_lines {
-                                let indicator = cursor_indicator(line_idx, current_line_idx);
-                                input_line.spans.insert(
-                                    0,
-                                    Span::styled(
-                                        indicator,
-                                        styles::current_line_indicator_style(&app.theme),
-                                    ),
-                                );
-                                lines.push(input_line);
-                                line_idx += 1;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Spacing between files
-        let indicator = cursor_indicator(line_idx, current_line_idx);
-        lines.push(Line::from(Span::styled(
-            indicator,
-            styles::current_line_indicator_style(&app.theme),
-        )));
-        line_idx += 1;
-    }
-
-    // Auto-scroll so the comment input box stays visible while the user types.
-    // Without this, adding a comment near the bottom/top of the viewport would
-    // place the input box off-screen and the user couldn't see what they type.
-    scroll_comment_input_into_view(
-        &mut app.diff_state.scroll_offset,
-        comment_input_box_range,
-        comment_cursor_logical_line,
-        inner.height as usize,
-        lines.len(),
-    );
-
-    let visible_lines_unscrolled: Vec<Line> = lines
-        .into_iter()
-        .skip(app.diff_state.scroll_offset)
-        .take(inner.height as usize)
-        .collect();
-
-    // Calculate the width of each line for max_content_width and visible line count
-    let line_widths: Vec<usize> = visible_lines_unscrolled
-        .iter()
-        .map(|line| {
-            line.spans
-                .iter()
-                .map(|span| span.content.width())
-                .sum::<usize>()
-        })
-        .collect();
-
-    let max_content_width = line_widths.iter().copied().max().unwrap_or(0);
-
-    app.diff_state.viewport_width = inner.width as usize;
-    app.diff_state.max_content_width = max_content_width;
-
-    let scroll_offset = app.diff_state.scroll_offset;
-    let wrap = app.diff_state.wrap_lines;
-    app.diff_state.visible_line_count = populate_row_to_annotation(
-        &mut app.diff_row_to_annotation,
-        &line_widths,
-        inner.width as usize,
-        inner.height as usize,
-        wrap,
-        scroll_offset,
-    );
-
-    let max_scroll_x = max_content_width.saturating_sub(inner.width as usize);
-    if app.diff_state.scroll_x > max_scroll_x {
-        app.diff_state.scroll_x = max_scroll_x;
-    }
-    if app.diff_state.wrap_lines {
-        app.diff_state.scroll_x = 0;
-    }
-
-    let scroll_x = app.diff_state.scroll_x;
-    let visible_lines_unscrolled_for_bg = visible_lines_unscrolled.clone();
-    let visible_lines: Vec<Line> = if app.diff_state.wrap_lines {
-        visible_lines_unscrolled
-    } else {
-        visible_lines_unscrolled
-            .into_iter()
-            .map(|line| apply_horizontal_scroll(line, scroll_x))
-            .collect()
-    };
-
-    // Paint per-visual-row add/del backgrounds across full row width.
-    paint_unified_diff_rows_with(
-        frame,
-        inner,
-        &visible_lines_unscrolled_for_bg,
-        &line_widths,
-        app.diff_state.wrap_lines,
-        inner.width as usize,
-        |_idx, line| unified_line_bg_style(line, &app.theme),
-    );
-
-    // Keep paragraph bg unset so pre-painted per-row diff backgrounds remain visible.
-    let mut diff = Paragraph::new(visible_lines).style(Style::default().fg(app.theme.fg_primary));
-    if app.diff_state.wrap_lines {
-        diff = diff.wrap(Wrap { trim: false });
-    }
-    frame.render_widget(diff, inner);
-
-    // Cursor-line bg has to land after the paragraph: spans on +/- lines carry
-    // explicit diff_add_bg/diff_del_bg that would mask a pre-paint over the code.
-    if app.cursor_line_highlight {
-        paint_unified_diff_rows_with(
-            frame,
-            inner,
-            &visible_lines_unscrolled_for_bg,
-            &line_widths,
-            app.diff_state.wrap_lines,
-            inner.width as usize,
-            |idx, _line| {
-                is_line_highlighted(app, idx).then(|| Style::default().bg(app.theme.cursor_line_bg))
-            },
-        );
-    }
-
-    if let Some(sel) = app.visual_selection {
-        paint_visual_selection_overlay(frame, inner, app, sel, &app.theme);
-    }
-
-    // Calculate screen position for comment cursor if in Comment mode
-    if let Some(cursor_logical_line) = comment_cursor_logical_line {
-        let scroll_offset = app.diff_state.scroll_offset;
-        // Use visible_line_count which accounts for line wrapping
-        let visible_lines_count = app.diff_state.visible_line_count.max(1);
-
-        // Check if the cursor line is visible (after scrolling)
-        if cursor_logical_line >= scroll_offset
-            && cursor_logical_line < scroll_offset + visible_lines_count
-        {
-            // Calculate screen row - need to account for wrapping
-            let logical_offset = cursor_logical_line - scroll_offset;
-
-            // Calculate visual row by summing wrapped line heights
-            let mut visual_row: u16 = 0;
-            let viewport_width = inner.width as usize;
-
-            if app.diff_state.wrap_lines && viewport_width > 0 {
-                // Calculate how many visual rows the lines before cursor take
-                // Note: line_widths is indexed from 0 and corresponds to visible lines
-                // (i.e., line_widths[0] is the first visible line after scroll)
-                for i in 0..logical_offset {
-                    if i < line_widths.len() {
-                        let width = line_widths[i];
-                        let rows = if width == 0 {
-                            1
-                        } else {
-                            width.div_ceil(viewport_width)
-                        };
-                        visual_row += rows as u16;
-                    } else {
-                        visual_row += 1;
-                    }
-                }
-            } else {
-                visual_row = logical_offset as u16;
-            }
-
-            // Account for diff area position (inner starts at diff block's inner area)
-            let screen_col = inner.x + comment_cursor_column;
-            let screen_row_abs = inner.y + visual_row;
-
-            app.comment_cursor_screen_pos = Some((screen_col, screen_row_abs));
-        }
-    }
-}
-
-/// Cursor info for the inline comment input box in side-by-side view:
-/// (cursor_logical_line, cursor_column, box_start_line, box_end_line)
-type SideBySideCursorInfo = (usize, u16, usize, usize, usize);
-
-/// Context for rendering side-by-side diff lines
-struct SideBySideContext<'a> {
-    app: &'a App,
-    theme: &'a Theme,
-    content_width: usize,
-    current_line_idx: usize,
-    // Comment input state for inline editing
-    comment_input_mode: bool,
-    comment_line: Option<(u32, LineSide)>,
-    comment_type: crate::model::CommentType,
-    comment_buffer: &'a str,
-    comment_cursor: usize,
-    comment_line_range: Option<LineRange>,
-    editing_comment_id: Option<&'a str>,
-    supports_keyboard_enhancement: bool,
-    current_file_idx: usize,
-}
-
-/// Get cursor indicator (single character for inline content)
-/// Populates `out` with the visual-row -> annotation-index map for the diff
-/// viewport and returns how many logical lines fit. Reuses the buffer's
-/// capacity to avoid per-frame allocations.
-fn populate_row_to_annotation(
-    out: &mut Vec<usize>,
-    line_widths: &[usize],
-    viewport_width: usize,
-    viewport_height: usize,
-    wrap: bool,
-    scroll_offset: usize,
-) -> usize {
-    out.clear();
-    out.reserve(viewport_height);
-    if wrap && viewport_width > 0 {
-        let mut visual_rows_used = 0;
-        let mut logical_lines_visible = 0;
-        for (i, &width) in line_widths.iter().enumerate() {
-            let rows_for_line = if width == 0 {
-                1
-            } else {
-                width.div_ceil(viewport_width)
-            };
-            if visual_rows_used + rows_for_line > viewport_height {
-                break;
-            }
-            for _ in 0..rows_for_line {
-                out.push(scroll_offset + i);
-            }
-            visual_rows_used += rows_for_line;
-            logical_lines_visible += 1;
-        }
-        logical_lines_visible.max(1)
-    } else {
-        for i in 0..line_widths.len().min(viewport_height) {
-            out.push(scroll_offset + i);
-        }
-        viewport_height
-    }
-}
-
-struct OverlayPaint {
-    sel: VisualSelection,
-    geom: crate::app::PaneGeom,
-    inner_left: u16,
-    inner_right: u16,
-    style: Style,
-}
-
-fn paint_visual_selection_overlay(
-    frame: &mut Frame,
-    inner: Rect,
-    app: &App,
-    sel: VisualSelection,
-    theme: &Theme,
-) {
-    let (start, end) = sel.ordered();
-    let paint = OverlayPaint {
-        sel,
-        geom: app.pane_geometry(inner, sel.anchor.side),
-        inner_left: inner.x,
-        inner_right: inner.x + inner.width.saturating_sub(1),
-        style: styles::visual_selection_style(theme),
-    };
-
-    let mut current: Option<(usize, u16, u16)> = None;
-    for rel in 0..app.diff_row_to_annotation.len() {
-        let ann_idx = app.diff_row_to_annotation[rel];
-        if ann_idx < start.annotation_idx {
-            continue;
-        }
-        if ann_idx > end.annotation_idx {
-            break;
-        }
-        let row = inner.y + rel as u16;
-        match current {
-            Some((cur, first, _)) if cur == ann_idx => {
-                current = Some((cur, first, row));
-            }
-            _ => {
-                if let Some(group) = current.take() {
-                    paint_annotation_group(frame, app, group, &paint);
-                }
-                current = Some((ann_idx, row, row));
-            }
-        }
-    }
-    if let Some(group) = current.take() {
-        paint_annotation_group(frame, app, group, &paint);
-    }
-}
-
-fn paint_annotation_group(
-    frame: &mut Frame,
-    app: &App,
-    group: (usize, u16, u16),
-    paint: &OverlayPaint,
-) {
-    let (ann_idx, first_row, last_row) = group;
-    if paint.geom.content_width == 0 {
-        return;
-    }
-
-    let side = paint.sel.anchor.side;
-    let group_height = (last_row - first_row) as usize + 1;
-    let pane_last_col = paint
-        .geom
-        .content_x_end
-        .saturating_sub(1)
-        .min(paint.inner_right);
-
-    let Some(content) = app.content_for_side(ann_idx, side) else {
-        // Headers and other non-content rows aren't bound by the pane
-        // gutter; tint the full inner width.
-        for which_row in 0..group_height {
-            let rect = Rect {
-                x: paint.inner_left,
-                y: first_row + which_row as u16,
-                width: paint.inner_right - paint.inner_left + 1,
-                height: 1,
-            };
-            frame.buffer_mut().set_style(rect, paint.style);
-        }
-        return;
-    };
-
-    let total_chars = content.chars().count();
-    let (lo, hi) = paint.sel.char_range(ann_idx, total_chars);
-    if hi <= lo {
-        return;
-    }
-
-    for which_row in 0..group_height {
-        let row_char_start = which_row * paint.geom.content_width;
-        let row_char_end = row_char_start + paint.geom.content_width;
-        let isect_lo = lo.max(row_char_start);
-        let isect_hi = hi.min(row_char_end);
-        if isect_hi <= isect_lo {
-            continue;
-        }
-        let col_lo_off = (isect_lo - row_char_start) as u16;
-        let col_hi_off = (isect_hi - row_char_start) as u16;
-        let col_lo = (paint.geom.content_x_start + col_lo_off).min(pane_last_col);
-        let col_hi_excl = paint.geom.content_x_start + col_hi_off;
-        if col_hi_excl == 0 {
-            continue;
-        }
-        let col_hi = col_hi_excl.saturating_sub(1).min(pane_last_col);
-        if col_lo > col_hi {
-            continue;
-        }
-        let rect = Rect {
-            x: col_lo,
-            y: first_row + which_row as u16,
-            width: col_hi - col_lo + 1,
-            height: 1,
-        };
-        frame.buffer_mut().set_style(rect, paint.style);
-    }
-}
-
-fn cursor_indicator(line_idx: usize, current_line_idx: usize) -> &'static str {
-    if line_idx == current_line_idx {
-        "▶"
-    } else {
-        " "
-    }
-}
-
-/// Get cursor indicator with spacing (two characters for line prefixes)
-fn cursor_indicator_spaced(line_idx: usize, current_line_idx: usize) -> &'static str {
-    if line_idx == current_line_idx {
-        "▶ "
-    } else {
-        "  "
-    }
-}
-
-/// Render a single expanded context line (shared by unified + side-by-side via unified path)
-fn render_expanded_context_line(
-    lines: &mut Vec<Line<'_>>,
-    line_idx: &mut usize,
-    current_line_idx: usize,
-    expanded_line: &crate::model::DiffLine,
-    theme: &Theme,
-) {
-    let indicator = cursor_indicator(*line_idx, current_line_idx);
-    let line_num = expanded_line
-        .new_lineno
-        .map(|n| format!("{n:>4} "))
-        .unwrap_or_else(|| "     ".to_string());
-    let line_spans = vec![
-        Span::styled(indicator, styles::current_line_indicator_style(theme)),
-        Span::styled(line_num, styles::expanded_context_style(theme)),
-        Span::styled("  ", styles::expanded_context_style(theme)),
-        Span::styled(
-            expanded_line.content.clone(),
-            styles::expanded_context_style(theme),
-        ),
-    ];
-    lines.push(Line::from(line_spans));
-    *line_idx += 1;
-}
-
-/// Render an expander line with direction arrow
-fn render_expander_line(
-    lines: &mut Vec<Line<'_>>,
-    line_idx: &mut usize,
-    current_line_idx: usize,
-    direction: ExpandDirection,
-    remaining: usize,
-    theme: &Theme,
-) {
-    let arrow = match direction {
-        ExpandDirection::Down => "↓",
-        ExpandDirection::Up => "↑",
-        ExpandDirection::Both => "↕",
-    };
-    let count = remaining.min(GAP_EXPAND_BATCH);
-    let indicator = cursor_indicator_spaced(*line_idx, current_line_idx);
-    lines.push(Line::from(vec![
-        Span::styled(indicator, styles::current_line_indicator_style(theme)),
-        Span::styled(
-            format!("       ... {arrow} expand ({count} lines) ..."),
-            styles::dim_style(theme),
-        ),
-    ]));
-    *line_idx += 1;
-}
-
-/// Render a single expanded context line in side-by-side mode
-fn render_sbs_expanded_context_line(
-    lines: &mut Vec<Line<'_>>,
-    line_idx: &mut usize,
-    current_line_idx: usize,
-    expanded_line: &crate::model::DiffLine,
-    content_width: usize,
-    theme: &Theme,
-) {
-    let indicator = cursor_indicator(*line_idx, current_line_idx);
-    let line_num = expanded_line
-        .new_lineno
-        .map(|n| format!("{n:>4} "))
-        .unwrap_or_else(|| "     ".to_string());
-    let line_spans = vec![
-        Span::styled(indicator, styles::current_line_indicator_style(theme)),
-        Span::styled(line_num.clone(), styles::expanded_context_style(theme)),
-        Span::styled(" ", styles::expanded_context_style(theme)),
-        Span::styled(
-            truncate_or_pad(&expanded_line.content, content_width),
-            styles::expanded_context_style(theme),
-        ),
-        Span::styled(" │ ", styles::dim_style(theme)),
-        Span::styled(line_num, styles::expanded_context_style(theme)),
-        Span::styled(" ", styles::expanded_context_style(theme)),
-        Span::styled(
-            truncate_or_pad(&expanded_line.content, content_width),
-            styles::expanded_context_style(theme),
-        ),
-    ];
-    lines.push(Line::from(line_spans));
-    *line_idx += 1;
-}
-
-/// Render a "N lines hidden" informational line
-fn render_hidden_lines(
-    lines: &mut Vec<Line<'_>>,
-    line_idx: &mut usize,
-    current_line_idx: usize,
-    count: usize,
-    theme: &Theme,
-) {
-    let indicator = cursor_indicator_spaced(*line_idx, current_line_idx);
-    lines.push(Line::from(vec![
-        Span::styled(indicator, styles::current_line_indicator_style(theme)),
-        Span::styled(
-            format!("       ... {count} lines hidden ..."),
-            styles::dim_style(theme),
-        ),
-    ]));
-    *line_idx += 1;
-}
-
-fn comment_type_presentation(
-    app: &App,
-    comment_type: &crate::model::CommentType,
-) -> comment_panel::CommentTypePresentation {
-    comment_panel::CommentTypePresentation {
-        label: app.comment_type_label(comment_type),
-        color: app.comment_type_color(comment_type),
-    }
-}
-
-/// Adjust scroll_offset so the comment input box is visible in the viewport.
-///
-/// The input box is rendered inline in the diff view, so without this
-/// adjustment it can end up below (or above) the visible area when a
-/// comment is started near the viewport edge or when typing a multi-line
-/// comment grows the box past the bottom. If the box is taller than the
-/// viewport we fall back to keeping just the text cursor line visible.
-fn scroll_comment_input_into_view(
-    scroll_offset: &mut usize,
-    box_range: Option<(usize, usize)>,
-    cursor_line: Option<usize>,
-    viewport_height: usize,
-    total_lines: usize,
-) {
-    let Some((box_start, box_end)) = box_range else {
-        return;
-    };
-    if viewport_height == 0 {
-        return;
-    }
-
-    let box_height = box_end.saturating_sub(box_start) + 1;
-
-    if box_height <= viewport_height {
-        if box_start < *scroll_offset {
-            *scroll_offset = box_start;
-        } else if box_end >= *scroll_offset + viewport_height {
-            *scroll_offset = box_end + 1 - viewport_height;
-        }
-    } else if let Some(cursor) = cursor_line {
-        // Box too tall for viewport: keep the text cursor line visible.
-        if cursor < *scroll_offset {
-            *scroll_offset = cursor;
-        } else if cursor >= *scroll_offset + viewport_height {
-            *scroll_offset = cursor + 1 - viewport_height;
-        }
-    }
-
-    // Clamp so we never scroll past the last line.
-    let max_scroll = total_lines.saturating_sub(viewport_height);
-    if *scroll_offset > max_scroll {
-        *scroll_offset = max_scroll;
-    }
-}
-
-fn render_side_by_side_diff(frame: &mut Frame, app: &mut App, area: Rect) {
-    let focused = app.focused_panel == FocusedPanel::Diff;
-
-    let title = if app.is_cursor_in_overview() || app.current_file_path().is_none() {
-        " Diff (Side-by-Side) \u{2014} Overview ".to_string()
-    } else {
-        format!(
-            " Diff (Side-by-Side) \u{2014} {} ",
-            app.current_file_path().unwrap().display()
-        )
-    };
-
-    let block = Block::default()
-        .title(title)
-        .title_top(diff_stat_title(app).right_aligned())
-        .borders(Borders::ALL)
-        .style(styles::panel_style(&app.theme))
-        .border_style(styles::border_style(&app.theme, focused));
-
-    let inner = block.inner(area);
-    frame.render_widget(block, area);
-
-    // Update viewport height for scroll calculations
-    app.diff_state.viewport_height = inner.height as usize;
-    app.diff_inner_area = Some(inner);
-
-    // Reset comment input annotation offset (will be set if a comment input box is rendered)
-    app.comment_input_annotation_offset = None;
-
-    // Layout: indicator(1) + linenum(4) + space(1) + prefix(1) + content + " │ "(3) + linenum(4) + space(1) + prefix(1) + content
-    let available_width = inner.width.saturating_sub(crate::app::SBS_OVERHEAD) as usize;
-    let content_width = available_width / 2;
-
-    // Determine if we're in line comment mode (not file-level)
-    let comment_input_mode = app.input_mode == InputMode::Comment
-        && !app.comment_is_file_level
-        && !app.comment_is_review_level;
-
-    let ctx = SideBySideContext {
-        app,
-        theme: &app.theme,
-        content_width,
-        current_line_idx: app.diff_state.cursor_line,
-        comment_input_mode,
-        comment_line: app.comment_line,
-        comment_type: app.comment_type.clone(),
-        comment_buffer: &app.comment_buffer,
-        comment_cursor: app.comment_cursor,
-        comment_line_range: app.comment_line_range.map(|(r, _)| r),
-        editing_comment_id: app.editing_comment_id.as_deref(),
-        supports_keyboard_enhancement: app.supports_keyboard_enhancement,
-        current_file_idx: app.diff_state.current_file_idx,
-    };
-
-    // Build all diff lines for side-by-side view
-    let mut lines: Vec<Line> = Vec::new();
-    let mut line_idx: usize = 0;
-
-    // Track cursor position for IME when in Comment mode
-    let mut comment_cursor_logical_line: Option<usize> = None;
-    let mut comment_cursor_column: u16 = 0;
-    // Track the full extent of the comment input box so we can auto-scroll
-    // the viewport to keep it visible while the user types.
-    let mut comment_input_box_range: Option<(usize, usize)> = None;
-    let mut annotation_offset: Option<(usize, usize, usize)> = None;
-
-    let is_review_comment_mode =
-        app.input_mode == InputMode::Comment && app.comment_is_review_level;
-
-    let general_indicator = cursor_indicator_spaced(line_idx, ctx.current_line_idx);
-    lines.push(Line::from(vec![
-        Span::styled(
-            general_indicator,
-            styles::current_line_indicator_style(&app.theme),
-        ),
-        Span::styled(
-            "═══ Review Comments ",
-            styles::file_header_style(&app.theme),
-        ),
-        Span::styled("═".repeat(40), styles::file_header_style(&app.theme)),
-    ]));
-    line_idx += 1;
-
-    for comment in &app.session.review_comments {
-        let is_being_edited =
-            app.editing_comment_id.as_ref() == Some(&comment.id) && is_review_comment_mode;
-
-        if is_being_edited {
-            let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-                &app.theme,
-                comment_type_presentation(app, &app.comment_type),
-                &app.comment_buffer,
-                app.comment_cursor,
-                None,
-                true,
-                app.supports_keyboard_enhancement,
-            );
-            comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-            comment_cursor_column = 1 + cursor_info.column;
-            comment_input_box_range =
-                Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-            let annotations_replaced = 2 + comment.content.split('\n').count();
-            annotation_offset = Some((line_idx, input_lines.len(), annotations_replaced));
-
-            for mut input_line in input_lines {
-                let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-                input_line.spans.insert(
-                    0,
-                    Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                );
-                lines.push(input_line);
-                line_idx += 1;
-            }
-        } else {
-            let comment_lines = comment_panel::format_comment_lines(
-                &app.theme,
-                comment_type_presentation(app, &comment.comment_type),
-                &comment.content,
-                None,
-            );
-            for mut comment_line in comment_lines {
-                let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-                comment_line.spans.insert(
-                    0,
-                    Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                );
-                lines.push(comment_line);
-                line_idx += 1;
-            }
-        }
-    }
-
-    if is_review_comment_mode && app.editing_comment_id.is_none() {
-        let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-            &app.theme,
-            comment_type_presentation(app, &app.comment_type),
-            &app.comment_buffer,
-            app.comment_cursor,
-            None,
-            false,
-            app.supports_keyboard_enhancement,
-        );
-        comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-        comment_cursor_column = 1 + cursor_info.column;
-        comment_input_box_range = Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-        annotation_offset = Some((line_idx, input_lines.len(), 0));
-
-        for mut input_line in input_lines {
-            let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-            input_line.spans.insert(
-                0,
-                Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-            );
-            lines.push(input_line);
-            line_idx += 1;
-        }
-    }
-
-    for (file_idx, file) in app.diff_files.iter().enumerate() {
-        let path = file.display_path();
-        let status = file.status.as_char();
-        let is_reviewed = app.session.is_file_reviewed(path);
-
-        // File header
-        let indicator = cursor_indicator_spaced(line_idx, ctx.current_line_idx);
-
-        let review_mark = if is_reviewed { "✓ " } else { "" };
-
-        let header_text = if file.is_commit_message {
-            format!("═══ {}Commit Message ", review_mark)
-        } else {
-            format!("═══ {}{} [{}] ", review_mark, path.display(), status)
-        };
-        lines.push(Line::from(vec![
-            Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-            Span::styled(header_text, styles::file_header_style(&app.theme)),
-            Span::styled("═".repeat(40), styles::file_header_style(&app.theme)),
-        ]));
-        line_idx += 1;
-
-        // If file is reviewed, skip rendering the body
-        if is_reviewed {
-            continue;
-        }
-
-        // Check if we're editing/adding a file-level comment for this file
-        let is_file_comment_mode = app.input_mode == InputMode::Comment
-            && app.comment_is_file_level
-            && file_idx == app.diff_state.current_file_idx;
-
-        // Show file-level comments
-        if let Some(review) = app.session.files.get(path) {
-            for comment in &review.file_comments {
-                // Skip rendering this comment if it's being edited
-                let is_being_edited =
-                    app.editing_comment_id.as_ref() == Some(&comment.id) && is_file_comment_mode;
-
-                if is_being_edited {
-                    // Render the inline input instead
-                    let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-                        &app.theme,
-                        comment_type_presentation(app, &app.comment_type),
-                        &app.comment_buffer,
-                        app.comment_cursor,
-                        None,
-                        true,
-                        app.supports_keyboard_enhancement,
-                    );
-                    comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-                    comment_cursor_column = 1 + cursor_info.column;
-                    comment_input_box_range =
-                        Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-                    let annotations_replaced = 2 + comment.content.split('\n').count();
-                    annotation_offset = Some((line_idx, input_lines.len(), annotations_replaced));
-
-                    for mut input_line in input_lines {
-                        let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-                        input_line.spans.insert(
-                            0,
-                            Span::styled(
-                                indicator,
-                                styles::current_line_indicator_style(&app.theme),
-                            ),
-                        );
-                        lines.push(input_line);
-                        line_idx += 1;
-                    }
-                } else {
-                    let comment_lines = comment_panel::format_comment_lines(
-                        &app.theme,
-                        comment_type_presentation(app, &comment.comment_type),
-                        &comment.content,
-                        None,
-                    );
-                    for mut comment_line in comment_lines {
-                        let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-                        comment_line.spans.insert(
-                            0,
-                            Span::styled(
-                                indicator,
-                                styles::current_line_indicator_style(&app.theme),
-                            ),
-                        );
-                        lines.push(comment_line);
-                        line_idx += 1;
-                    }
-                }
-            }
-        }
-
-        // Render inline input for new file-level comment
-        if is_file_comment_mode && app.editing_comment_id.is_none() {
-            let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-                &app.theme,
-                comment_type_presentation(app, &app.comment_type),
-                &app.comment_buffer,
-                app.comment_cursor,
-                None,
-                false,
-                app.supports_keyboard_enhancement,
-            );
-            comment_cursor_logical_line = Some(line_idx + cursor_info.line_offset);
-            comment_cursor_column = 1 + cursor_info.column;
-            comment_input_box_range =
-                Some((line_idx, line_idx + input_lines.len().saturating_sub(1)));
-            annotation_offset = Some((line_idx, input_lines.len(), 0));
-
-            for mut input_line in input_lines {
-                let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-                input_line.spans.insert(
-                    0,
-                    Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                );
-                lines.push(input_line);
-                line_idx += 1;
-            }
-        }
-
-        if file.is_too_large {
-            let indicator = cursor_indicator_spaced(line_idx, ctx.current_line_idx);
-            lines.push(Line::from(vec![
-                Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                Span::styled("(file too large to display)", styles::dim_style(&app.theme)),
-            ]));
-            line_idx += 1;
-        } else if file.is_binary {
-            let indicator = cursor_indicator_spaced(line_idx, ctx.current_line_idx);
-            lines.push(Line::from(vec![
-                Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                Span::styled("(binary file)", styles::dim_style(&app.theme)),
-            ]));
-            line_idx += 1;
-        } else if file.hunks.is_empty() {
-            let indicator = cursor_indicator_spaced(line_idx, ctx.current_line_idx);
-            lines.push(Line::from(vec![
-                Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                Span::styled("(no changes)", styles::dim_style(&app.theme)),
-            ]));
-            line_idx += 1;
-        } else {
-            let line_comments = app
-                .session
-                .files
-                .get(path)
-                .map(|r| &r.line_comments)
-                .cloned()
-                .unwrap_or_default();
-
-            for (hunk_idx, hunk) in file.hunks.iter().enumerate() {
-                // Calculate and render gap before this hunk
-                let prev_hunk = if hunk_idx > 0 {
-                    file.hunks.get(hunk_idx - 1)
-                } else {
-                    None
-                };
-                let gap = calculate_gap(
-                    prev_hunk.map(|h| (&h.new_start, &h.new_count)),
-                    hunk.new_start,
-                );
-
-                let gap_id = GapId { file_idx, hunk_idx };
-
-                if gap > 0 {
-                    let top_lines = app.expanded_top.get(&gap_id);
-                    let bot_lines = app.expanded_bottom.get(&gap_id);
-                    let top_len = top_lines.map_or(0, |v| v.len());
-                    let bot_len = bot_lines.map_or(0, |v| v.len());
-                    let remaining = (gap as usize).saturating_sub(top_len + bot_len);
-                    let is_top_of_file = hunk_idx == 0;
-
-                    // Render top expanded lines
-                    if let Some(top) = top_lines {
-                        for expanded_line in top {
-                            render_sbs_expanded_context_line(
-                                &mut lines,
-                                &mut line_idx,
-                                ctx.current_line_idx,
-                                expanded_line,
-                                ctx.content_width,
-                                &app.theme,
-                            );
-                        }
-                    }
-
-                    // Render expanders / hidden lines
-                    if remaining > 0 {
-                        if is_top_of_file {
-                            if remaining > GAP_EXPAND_BATCH {
-                                render_hidden_lines(
-                                    &mut lines,
-                                    &mut line_idx,
-                                    ctx.current_line_idx,
-                                    remaining,
-                                    &app.theme,
-                                );
-                            }
-                            render_expander_line(
-                                &mut lines,
-                                &mut line_idx,
-                                ctx.current_line_idx,
-                                ExpandDirection::Up,
-                                remaining,
-                                &app.theme,
-                            );
-                        } else if remaining >= GAP_EXPAND_BATCH {
-                            render_expander_line(
-                                &mut lines,
-                                &mut line_idx,
-                                ctx.current_line_idx,
-                                ExpandDirection::Down,
-                                remaining,
-                                &app.theme,
-                            );
-                            render_hidden_lines(
-                                &mut lines,
-                                &mut line_idx,
-                                ctx.current_line_idx,
-                                remaining,
-                                &app.theme,
-                            );
-                            render_expander_line(
-                                &mut lines,
-                                &mut line_idx,
-                                ctx.current_line_idx,
-                                ExpandDirection::Up,
-                                remaining,
-                                &app.theme,
-                            );
-                        } else {
-                            render_expander_line(
-                                &mut lines,
-                                &mut line_idx,
-                                ctx.current_line_idx,
-                                ExpandDirection::Both,
-                                remaining,
-                                &app.theme,
-                            );
-                        }
-                    }
-
-                    // Render bottom expanded lines
-                    if let Some(bot) = bot_lines {
-                        for expanded_line in bot {
-                            render_sbs_expanded_context_line(
-                                &mut lines,
-                                &mut line_idx,
-                                ctx.current_line_idx,
-                                expanded_line,
-                                ctx.content_width,
-                                &app.theme,
-                            );
-                        }
-                    }
-                }
-
-                // Hunk header
-                let indicator = cursor_indicator_spaced(line_idx, ctx.current_line_idx);
-                lines.push(Line::from(vec![
-                    Span::styled(indicator, styles::current_line_indicator_style(&app.theme)),
-                    Span::styled(
-                        hunk.header.to_string(),
-                        styles::diff_hunk_header_style(&app.theme),
-                    ),
-                ]));
-                line_idx += 1;
-
-                // Process diff lines in side-by-side format
-                let (new_line_idx, cursor_info) = render_hunk_lines_side_by_side(
-                    &hunk.lines,
-                    &line_comments,
-                    &ctx,
-                    file_idx,
-                    line_idx,
-                    &mut lines,
-                );
-                line_idx = new_line_idx;
-                if let Some((line, col, box_start, box_end, annotations_replaced)) = cursor_info {
-                    comment_cursor_logical_line = Some(line);
-                    comment_cursor_column = col;
-                    comment_input_box_range = Some((box_start, box_end));
-                    let box_len = box_end - box_start + 1;
-                    annotation_offset = Some((box_start, box_len, annotations_replaced));
-                }
-            }
-        }
-
-        // Spacing between files
-        let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-        lines.push(Line::from(Span::styled(
-            indicator,
-            styles::current_line_indicator_style(&app.theme),
-        )));
-        line_idx += 1;
-    }
-
-    drop(ctx);
-    app.comment_input_annotation_offset = annotation_offset;
-
-    // Auto-scroll so the comment input box stays visible while the user types.
-    scroll_comment_input_into_view(
-        &mut app.diff_state.scroll_offset,
-        comment_input_box_range,
-        comment_cursor_logical_line,
-        inner.height as usize,
-        lines.len(),
-    );
-
-    let visible_lines_unscrolled: Vec<Line> = lines
-        .into_iter()
-        .skip(app.diff_state.scroll_offset)
-        .take(inner.height as usize)
-        .collect();
-
-    // Calculate the width of each line for max_content_width and visible line count
-    let line_widths: Vec<usize> = visible_lines_unscrolled
-        .iter()
-        .map(|line| {
-            line.spans
-                .iter()
-                .map(|span| span.content.width())
-                .sum::<usize>()
-        })
-        .collect();
-
-    let max_content_width = line_widths.iter().copied().max().unwrap_or(0);
-
-    app.diff_state.viewport_width = inner.width as usize;
-    app.diff_state.max_content_width = max_content_width;
-
-    let scroll_offset = app.diff_state.scroll_offset;
-    let wrap = app.diff_state.wrap_lines;
-    app.diff_state.visible_line_count = populate_row_to_annotation(
-        &mut app.diff_row_to_annotation,
-        &line_widths,
-        inner.width as usize,
-        inner.height as usize,
-        wrap,
-        scroll_offset,
-    );
-
-    let max_scroll_x = max_content_width.saturating_sub(inner.width as usize);
-    if app.diff_state.scroll_x > max_scroll_x {
-        app.diff_state.scroll_x = max_scroll_x;
-    }
-    if app.diff_state.wrap_lines {
-        app.diff_state.scroll_x = 0;
-    }
-
-    let scroll_x = app.diff_state.scroll_x;
-    let visible_lines: Vec<Line> = if app.diff_state.wrap_lines {
-        visible_lines_unscrolled
-    } else {
-        visible_lines_unscrolled
-            .into_iter()
-            .map(|line| apply_horizontal_scroll(line, scroll_x))
-            .collect()
-    };
-
-    let mut diff = Paragraph::new(visible_lines).style(styles::panel_style(&app.theme));
-    if app.diff_state.wrap_lines {
-        diff = diff.wrap(Wrap { trim: false });
-    }
-    frame.render_widget(diff, inner);
-
-    if app.cursor_line_highlight {
-        let viewport_height = inner.height as usize;
-        for offset in 0..viewport_height {
-            if is_line_highlighted(app, offset) {
-                let row_rect = Rect {
-                    x: inner.x,
-                    y: inner.y + offset as u16,
-                    width: inner.width,
-                    height: 1,
-                };
-                frame
-                    .buffer_mut()
-                    .set_style(row_rect, Style::default().bg(app.theme.cursor_line_bg));
-            }
-        }
-    }
-
-    // Painted last so the cell overlay wins over cursor-line bg on overlap.
-    if let Some(sel) = app.visual_selection {
-        paint_visual_selection_overlay(frame, inner, app, sel, &app.theme);
-    }
-
-    // Calculate screen position for comment cursor if in Comment mode
-    if let Some(cursor_logical_line) = comment_cursor_logical_line {
-        let scroll_offset = app.diff_state.scroll_offset;
-        let visible_lines_count = app.diff_state.visible_line_count.max(1);
-
-        // Check if the cursor line is visible (after scrolling)
-        if cursor_logical_line >= scroll_offset
-            && cursor_logical_line < scroll_offset + visible_lines_count
-        {
-            // Calculate screen row - need to account for wrapping
-            let logical_offset = cursor_logical_line - scroll_offset;
-
-            let mut visual_row: u16 = 0;
-            let viewport_width = inner.width as usize;
-
-            if app.diff_state.wrap_lines && viewport_width > 0 {
-                for i in 0..logical_offset {
-                    if i < line_widths.len() {
-                        let width = line_widths[i];
-                        let rows = if width == 0 {
-                            1
-                        } else {
-                            width.div_ceil(viewport_width)
-                        };
-                        visual_row += rows as u16;
-                    } else {
-                        visual_row += 1;
-                    }
-                }
-            } else {
-                visual_row = logical_offset as u16;
-            }
-
-            let screen_col = inner.x + comment_cursor_column;
-            let screen_row_abs = inner.y + visual_row;
-
-            app.comment_cursor_screen_pos = Some((screen_col, screen_row_abs));
-        }
-    }
-}
-
-/// Process and render all diff lines in a hunk for side-by-side view
-/// Returns (new_line_idx, optional cursor info for inline comment input)
-fn render_hunk_lines_side_by_side(
-    hunk_lines: &[crate::model::DiffLine],
-    line_comments: &std::collections::HashMap<u32, Vec<crate::model::Comment>>,
-    ctx: &SideBySideContext,
-    file_idx: usize,
-    mut line_idx: usize,
-    lines: &mut Vec<Line>,
-) -> (usize, Option<SideBySideCursorInfo>) {
-    let mut i = 0;
-    let mut cursor_info_out: Option<SideBySideCursorInfo> = None;
-
-    while i < hunk_lines.len() {
-        let diff_line = &hunk_lines[i];
-
-        match diff_line.origin {
-            LineOrigin::Context => {
-                let (new_line_idx, cursor_info) = render_context_line_side_by_side(
-                    diff_line,
-                    line_comments,
-                    ctx,
-                    file_idx,
-                    line_idx,
-                    lines,
-                );
-                line_idx = new_line_idx;
-                if cursor_info.is_some() {
-                    cursor_info_out = cursor_info;
-                }
-                i += 1;
-            }
-            LineOrigin::Deletion => {
-                let (new_line_idx, lines_processed, cursor_info) =
-                    render_deletion_addition_pair_side_by_side(
-                        hunk_lines,
-                        i,
-                        line_comments,
-                        ctx,
-                        file_idx,
-                        line_idx,
-                        lines,
-                    );
-                line_idx = new_line_idx;
-                if cursor_info.is_some() {
-                    cursor_info_out = cursor_info;
-                }
-                i = lines_processed;
-            }
-            LineOrigin::Addition => {
-                let (new_line_idx, cursor_info) = render_standalone_addition_side_by_side(
-                    diff_line,
-                    line_comments,
-                    ctx,
-                    file_idx,
-                    line_idx,
-                    lines,
-                );
-                line_idx = new_line_idx;
-                if cursor_info.is_some() {
-                    cursor_info_out = cursor_info;
-                }
-                i += 1;
-            }
-        }
-    }
-    (line_idx, cursor_info_out)
-}
-
-/// Render a context line (appears on both sides)
-/// Returns (new_line_idx, optional cursor info for inline comment input)
-fn render_context_line_side_by_side(
-    diff_line: &crate::model::DiffLine,
-    line_comments: &std::collections::HashMap<u32, Vec<crate::model::Comment>>,
-    ctx: &SideBySideContext,
-    file_idx: usize,
-    mut line_idx: usize,
-    lines: &mut Vec<Line>,
-) -> (usize, Option<SideBySideCursorInfo>) {
-    let line_num = diff_line
-        .old_lineno
-        .or(diff_line.new_lineno)
-        .map(|n| format!("{n:>4}"))
-        .unwrap_or_else(|| "    ".to_string());
-
-    let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-
-    let mut spans = vec![
-        Span::styled(indicator, styles::current_line_indicator_style(ctx.theme)),
-        Span::styled(format!("{line_num} "), styles::dim_style(ctx.theme)),
-        Span::styled(" ".to_string(), styles::diff_context_style(ctx.theme)),
-    ];
-
-    // Left side content - use syntax highlighting if available
-    if let Some(ref highlighted) = diff_line.highlighted_spans {
-        let content_spans = truncate_or_pad_spans(
-            highlighted,
-            ctx.content_width,
-            styles::diff_context_style(ctx.theme),
-        );
-        spans.extend(content_spans);
-    } else {
-        let content = truncate_or_pad(&diff_line.content, ctx.content_width);
-        spans.push(Span::styled(content, styles::diff_context_style(ctx.theme)));
-    }
-
-    // Separator
-    spans.push(Span::styled(" │ ", styles::dim_style(ctx.theme)));
-    spans.push(Span::styled(
-        format!("{line_num} "),
-        styles::dim_style(ctx.theme),
-    ));
-    spans.push(Span::styled(
-        " ".to_string(),
-        styles::diff_context_style(ctx.theme),
-    ));
-
-    // Right side content - use same highlighting
-    if let Some(ref highlighted) = diff_line.highlighted_spans {
-        let content_spans = truncate_or_pad_spans(
-            highlighted,
-            ctx.content_width,
-            styles::diff_context_style(ctx.theme),
-        );
-        spans.extend(content_spans);
-    } else {
-        let content = truncate_or_pad(&diff_line.content, ctx.content_width);
-        spans.push(Span::styled(content, styles::diff_context_style(ctx.theme)));
-    }
-
-    lines.push(Line::from(spans));
-    line_idx += 1;
-
-    // Add comments if any
-    let mut cursor_info_out: Option<SideBySideCursorInfo> = None;
-    if let Some(new_ln) = diff_line.new_lineno {
-        let (new_line_idx, cursor_info) = add_comments_to_line(
-            new_ln,
-            line_comments,
-            LineSide::New,
-            ctx,
-            file_idx,
-            line_idx,
-            lines,
-        );
-        line_idx = new_line_idx;
-        cursor_info_out = cursor_info;
-    }
-
-    (line_idx, cursor_info_out)
-}
-
-/// Render paired deletions and additions side-by-side
-/// Returns (line_idx, skip_count, optional cursor info for inline comment input)
-fn render_deletion_addition_pair_side_by_side(
-    hunk_lines: &[crate::model::DiffLine],
-    start_idx: usize,
-    line_comments: &std::collections::HashMap<u32, Vec<crate::model::Comment>>,
-    ctx: &SideBySideContext,
-    file_idx: usize,
-    mut line_idx: usize,
-    lines: &mut Vec<Line>,
-) -> (usize, usize, Option<SideBySideCursorInfo>) {
-    // Find the range of consecutive deletions
-    let mut del_end = start_idx + 1;
-    while del_end < hunk_lines.len() && hunk_lines[del_end].origin == LineOrigin::Deletion {
-        del_end += 1;
-    }
-
-    // Find the range of consecutive additions following the deletions
-    let add_start = del_end;
-    let mut add_end = add_start;
-    while add_end < hunk_lines.len() && hunk_lines[add_end].origin == LineOrigin::Addition {
-        add_end += 1;
-    }
-
-    let del_count = del_end - start_idx;
-    let add_count = add_end - add_start;
-    let max_lines = del_count.max(add_count);
-    let mut cursor_info_out: Option<SideBySideCursorInfo> = None;
-
-    // Render each pair of deletion/addition
-    for offset in 0..max_lines {
-        let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-
-        let mut spans = vec![Span::styled(
-            indicator,
-            styles::current_line_indicator_style(ctx.theme),
-        )];
-
-        // Left side (deletion)
-        if offset < del_count {
-            let del_line = &hunk_lines[start_idx + offset];
-            add_deletion_spans(ctx.theme, &mut spans, del_line, ctx.content_width);
-        } else {
-            add_empty_column_spans(&mut spans, ctx.content_width);
-        }
-
-        spans.push(Span::styled(" │ ", styles::dim_style(ctx.theme)));
-
-        // Right side (addition)
-        if offset < add_count {
-            let add_line = &hunk_lines[add_start + offset];
-            add_addition_spans(ctx.theme, &mut spans, add_line, ctx.content_width);
-        } else {
-            add_empty_column_spans(&mut spans, ctx.content_width);
-        }
-
-        lines.push(Line::from(spans));
-        line_idx += 1;
-
-        // Add comments for deletion
-        if offset < del_count {
-            let del_line = &hunk_lines[start_idx + offset];
-            if let Some(old_ln) = del_line.old_lineno {
-                let (new_line_idx, cursor_info) = add_comments_to_line(
-                    old_ln,
-                    line_comments,
-                    LineSide::Old,
-                    ctx,
-                    file_idx,
-                    line_idx,
-                    lines,
-                );
-                line_idx = new_line_idx;
-                if cursor_info.is_some() {
-                    cursor_info_out = cursor_info;
-                }
-            }
-        }
-
-        // Add comments for addition
-        if offset < add_count {
-            let add_line = &hunk_lines[add_start + offset];
-            if let Some(new_ln) = add_line.new_lineno {
-                let (new_line_idx, cursor_info) = add_comments_to_line(
-                    new_ln,
-                    line_comments,
-                    LineSide::New,
-                    ctx,
-                    file_idx,
-                    line_idx,
-                    lines,
-                );
-                line_idx = new_line_idx;
-                if cursor_info.is_some() {
-                    cursor_info_out = cursor_info;
-                }
-            }
-        }
-    }
-
-    (line_idx, add_end, cursor_info_out)
-}
-
-/// Render a standalone addition (no matching deletion)
-/// Returns (new_line_idx, optional cursor info for inline comment input)
-fn render_standalone_addition_side_by_side(
-    diff_line: &crate::model::DiffLine,
-    line_comments: &std::collections::HashMap<u32, Vec<crate::model::Comment>>,
-    ctx: &SideBySideContext,
-    file_idx: usize,
-    mut line_idx: usize,
-    lines: &mut Vec<Line>,
-) -> (usize, Option<SideBySideCursorInfo>) {
-    let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-
-    let mut spans = vec![Span::styled(
-        indicator,
-        styles::current_line_indicator_style(ctx.theme),
-    )];
-    add_empty_column_spans(&mut spans, ctx.content_width);
-    spans.push(Span::styled(" │ ", styles::dim_style(ctx.theme)));
-    add_addition_spans(ctx.theme, &mut spans, diff_line, ctx.content_width);
-
-    lines.push(Line::from(spans));
-    line_idx += 1;
-
-    // Add comments if any
-    let mut cursor_info_out: Option<SideBySideCursorInfo> = None;
-    if let Some(new_ln) = diff_line.new_lineno {
-        let (new_line_idx, cursor_info) = add_comments_to_line(
-            new_ln,
-            line_comments,
-            LineSide::New,
-            ctx,
-            file_idx,
-            line_idx,
-            lines,
-        );
-        line_idx = new_line_idx;
-        cursor_info_out = cursor_info;
-    }
-
-    (line_idx, cursor_info_out)
-}
-
-/// Add deletion line spans to the spans vector
-fn add_deletion_spans(
-    theme: &Theme,
-    spans: &mut Vec<Span>,
-    diff_line: &crate::model::DiffLine,
-    content_width: usize,
-) {
-    let line_num = diff_line
-        .old_lineno
-        .map(|n| format!("{n:>4}"))
-        .unwrap_or_else(|| "    ".to_string());
-
-    spans.push(Span::styled(
-        format!("{line_num} "),
-        styles::dim_style(theme),
-    ));
-    spans.push(Span::styled("-".to_string(), styles::diff_del_style(theme)));
-
-    // Use syntax highlighting if available
-    if let Some(ref highlighted) = diff_line.highlighted_spans {
-        let syntax_pad_style = Style::default().fg(theme.diff_del).bg(theme.syntax_del_bg);
-        let content_spans = truncate_or_pad_spans(highlighted, content_width, syntax_pad_style);
-        spans.extend(content_spans);
-    } else {
-        // Fall back to plain text
-        let content = truncate_or_pad(&diff_line.content, content_width);
-        spans.push(Span::styled(content, styles::diff_del_style(theme)));
-    }
-}
-
-/// Add addition line spans to the spans vector
-fn add_addition_spans(
-    theme: &Theme,
-    spans: &mut Vec<Span>,
-    diff_line: &crate::model::DiffLine,
-    content_width: usize,
-) {
-    let line_num = diff_line
-        .new_lineno
-        .map(|n| format!("{n:>4}"))
-        .unwrap_or_else(|| "    ".to_string());
-
-    spans.push(Span::styled(
-        format!("{line_num} "),
-        styles::dim_style(theme),
-    ));
-    spans.push(Span::styled("+".to_string(), styles::diff_add_style(theme)));
-
-    // Use syntax highlighting if available
-    if let Some(ref highlighted) = diff_line.highlighted_spans {
-        let syntax_pad_style = Style::default().fg(theme.diff_add).bg(theme.syntax_add_bg);
-        let content_spans = truncate_or_pad_spans(highlighted, content_width, syntax_pad_style);
-        spans.extend(content_spans);
-    } else {
-        // Fall back to plain text
-        let content = truncate_or_pad(&diff_line.content, content_width);
-        spans.push(Span::styled(content, styles::diff_add_style(theme)));
-    }
-}
-
-/// Add empty column spans (for when one side has no content)
-fn add_empty_column_spans(spans: &mut Vec<Span>, content_width: usize) {
-    // line_num(4) + space(1) + prefix(1) + content
-    spans.push(Span::styled(
-        " ".repeat(5 + 1 + content_width),
-        Style::default(),
-    ));
-}
-
-/// Add comments for a specific line.
-/// Returns (new_line_idx, optional cursor info for inline comment input)
-fn add_comments_to_line(
-    line_num: u32,
-    line_comments: &std::collections::HashMap<u32, Vec<crate::model::Comment>>,
-    side: LineSide,
-    ctx: &SideBySideContext,
-    file_idx: usize,
-    mut line_idx: usize,
-    lines: &mut Vec<Line>,
-) -> (usize, Option<SideBySideCursorInfo>) {
-    // Check if we're adding/editing a comment on this line and side
-    let is_line_comment_mode = ctx.comment_input_mode
-        && file_idx == ctx.current_file_idx
-        && ctx.comment_line == Some((line_num, side));
-    let mut cursor_info_out: Option<SideBySideCursorInfo> = None;
-
-    if let Some(comments) = line_comments.get(&line_num) {
-        for comment in comments {
-            let comment_side = comment.side.unwrap_or(LineSide::New);
-            if (side == LineSide::Old && comment_side == LineSide::Old)
-                || (side == LineSide::New && comment_side != LineSide::Old)
-            {
-                // Check if this comment is being edited
-                let is_being_edited =
-                    is_line_comment_mode && ctx.editing_comment_id == Some(comment.id.as_str());
-
-                if is_being_edited {
-                    // Render inline input instead
-                    let line_range = ctx
-                        .comment_line_range
-                        .or_else(|| Some(LineRange::single(line_num)));
-                    let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-                        ctx.theme,
-                        comment_type_presentation(ctx.app, &ctx.comment_type),
-                        ctx.comment_buffer,
-                        ctx.comment_cursor,
-                        line_range,
-                        true,
-                        ctx.supports_keyboard_enhancement,
-                    );
-                    let box_end = line_idx + input_lines.len().saturating_sub(1);
-                    let annotations_replaced = 2 + comment.content.split('\n').count();
-                    cursor_info_out = Some((
-                        line_idx + cursor_info.line_offset,
-                        1 + cursor_info.column,
-                        line_idx,
-                        box_end,
-                        annotations_replaced,
-                    ));
-
-                    for mut input_line in input_lines {
-                        let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-                        input_line.spans.insert(
-                            0,
-                            Span::styled(
-                                indicator,
-                                styles::current_line_indicator_style(ctx.theme),
-                            ),
-                        );
-                        lines.push(input_line);
-                        line_idx += 1;
-                    }
-                } else {
-                    let line_range = comment
-                        .line_range
-                        .or_else(|| Some(LineRange::single(line_num)));
-                    let comment_lines = comment_panel::format_comment_lines(
-                        ctx.theme,
-                        comment_type_presentation(ctx.app, &comment.comment_type),
-                        &comment.content,
-                        line_range,
-                    );
-                    for mut comment_line in comment_lines {
-                        let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-                        comment_line.spans.insert(
-                            0,
-                            Span::styled(
-                                indicator,
-                                styles::current_line_indicator_style(ctx.theme),
-                            ),
-                        );
-                        lines.push(comment_line);
-                        line_idx += 1;
-                    }
-                }
-            }
-        }
-    }
-
-    // Render inline input for new line comment
-    if is_line_comment_mode && ctx.editing_comment_id.is_none() {
-        let line_range = ctx
-            .comment_line_range
-            .or_else(|| Some(LineRange::single(line_num)));
-        let (input_lines, cursor_info) = comment_panel::format_comment_input_lines(
-            ctx.theme,
-            comment_type_presentation(ctx.app, &ctx.comment_type),
-            ctx.comment_buffer,
-            ctx.comment_cursor,
-            line_range,
-            false,
-            ctx.supports_keyboard_enhancement,
-        );
-        let box_end = line_idx + input_lines.len().saturating_sub(1);
-        cursor_info_out = Some((
-            line_idx + cursor_info.line_offset,
-            1 + cursor_info.column,
-            line_idx,
-            box_end,
-            0,
-        ));
-
-        for mut input_line in input_lines {
-            let indicator = cursor_indicator(line_idx, ctx.current_line_idx);
-            input_line.spans.insert(
-                0,
-                Span::styled(indicator, styles::current_line_indicator_style(ctx.theme)),
-            );
-            lines.push(input_line);
-            line_idx += 1;
-        }
-    }
-
-    (line_idx, cursor_info_out)
-}
-
-/// Truncate or pad a string to a specific width
-fn truncate_or_pad(s: &str, width: usize) -> String {
-    let char_count = s.chars().count();
-    if char_count > width {
-        s.chars().take(width.saturating_sub(3)).collect::<String>() + "..."
-    } else {
-        format!("{s:width$}")
-    }
-}
-
-/// Truncate or pad highlighted spans to a specific display width
-/// Uses unicode width to properly handle wide characters (CJK, emoji, etc.)
-/// Returns a vector of spans that fits exactly within the width
-fn truncate_or_pad_spans(
-    spans: &[(Style, String)],
-    width: usize,
-    base_style: Style,
-) -> Vec<Span<'static>> {
-    // Count total display width
-    let total_width: usize = spans.iter().map(|(_, text)| text.width()).sum();
-
-    if total_width > width {
-        // Need to truncate
-        let mut result = Vec::new();
-        let mut remaining = width.saturating_sub(3); // Reserve space for "..."
-
-        for (style, text) in spans {
-            if remaining == 0 {
-                break;
-            }
-
-            let text_width = text.width();
-            if text_width <= remaining {
-                result.push(Span::styled(text.clone(), *style));
-                remaining -= text_width;
-            } else {
-                // Truncate this span character by character to fit remaining width
-                let mut truncated = String::new();
-                let mut current_width = 0;
-                for c in text.chars() {
-                    let char_width = unicode_width::UnicodeWidthChar::width(c).unwrap_or(0);
-                    if current_width + char_width > remaining {
-                        break;
-                    }
-                    truncated.push(c);
-                    current_width += char_width;
-                }
-                if !truncated.is_empty() {
-                    result.push(Span::styled(truncated, *style));
-                }
-                remaining = 0;
-            }
-        }
-
-        // Add ellipsis
-        result.push(Span::styled("...".to_string(), base_style));
-        result
-    } else if total_width < width {
-        // Need to pad
-        let mut result: Vec<Span> = spans
-            .iter()
-            .map(|(style, text)| Span::styled(text.clone(), *style))
-            .collect();
-
-        // Add padding
-        let padding = " ".repeat(width - total_width);
-        result.push(Span::styled(padding, base_style));
-        result
-    } else {
-        // Perfect fit
-        spans
-            .iter()
-            .map(|(style, text)| Span::styled(text.clone(), *style))
-            .collect()
-    }
-}
-
-fn is_line_highlighted(app: &App, viewport_idx: usize) -> bool {
-    if !app.cursor_line_highlight {
-        return false;
-    }
-
-    let abs_idx = viewport_idx + app.diff_state.scroll_offset;
-
-    // Cursor line
-    if abs_idx == app.diff_state.cursor_line {
-        return true;
-    }
-
-    // Carryover from V → c: keep the comment-input box lit. The visual
-    // selection itself paints via the cell-precise overlay.
-    let Some((range, sel_side)) = app.comment_line_range else {
-        return false;
-    };
-
-    // Adjust the annotation index to account for the comment input box, which
-    // may have a different line count than what line_annotations expects.
-    let annotation_idx =
-        if let Some((box_start, box_len, replaced)) = app.comment_input_annotation_offset {
-            if abs_idx < box_start {
-                abs_idx
-            } else if abs_idx < box_start + box_len {
-                // Inside the comment input box - only highlight the portion that
-                // maps to annotation entries being replaced (edited comment lines)
-                let offset_in_box = abs_idx - box_start;
-                if offset_in_box < replaced {
-                    box_start + offset_in_box
-                } else {
-                    return false;
-                }
-            } else {
-                // After the box: shift by the difference between rendered and annotation counts
-                // box_len > replaced: input box added extra lines → shift back
-                // box_len < replaced: input box is shorter → shift forward
-                abs_idx + replaced - box_len
-            }
-        } else {
-            abs_idx
-        };
-
-    let Some(annotation) = app.line_annotations.get(annotation_idx) else {
-        return false;
-    };
-    let (file_idx, lineno) = match annotation {
-        AnnotatedLine::DiffLine {
-            file_idx,
-            old_lineno,
-            new_lineno,
-            ..
-        }
-        | AnnotatedLine::SideBySideLine {
-            file_idx,
-            old_lineno,
-            new_lineno,
-            ..
-        } => {
-            let ln = match sel_side {
-                LineSide::New => *new_lineno,
-                LineSide::Old => *old_lineno,
-            };
-            (*file_idx, ln)
-        }
-        _ => return false,
-    };
-    file_idx == app.diff_state.current_file_idx && lineno.is_some_and(|ln| range.contains(ln))
-}
-
-fn unified_line_bg_style(line: &Line, theme: &Theme) -> Option<Style> {
-    let prefix = line.spans.get(2)?.content.as_ref();
-    let default_bg = match prefix {
-        "+ " => theme.diff_add_bg,
-        "- " => theme.diff_del_bg,
-        _ => return None,
-    };
-
-    let bg = line
-        .spans
-        .last()
-        .and_then(|span| span.style.bg)
-        .unwrap_or(default_bg);
-
-    Some(Style::default().bg(bg))
-}
-
-fn paint_unified_diff_rows_with<F>(
-    frame: &mut Frame,
-    inner: Rect,
-    visible_lines_unscrolled: &[Line],
-    line_widths: &[usize],
-    wrap_lines: bool,
-    viewport_width: usize,
-    style_for: F,
-) where
-    F: Fn(usize, &Line) -> Option<Style>,
-{
-    let mut visual_row: usize = 0;
-
-    for (idx, line) in visible_lines_unscrolled.iter().enumerate() {
-        if visual_row >= inner.height as usize {
-            break;
-        }
-
-        let rows_for_line = if wrap_lines && viewport_width > 0 {
-            let width = line_widths.get(idx).copied().unwrap_or(0);
-            if width == 0 {
-                1
-            } else {
-                width.div_ceil(viewport_width)
-            }
-        } else {
-            1
-        };
-
-        if let Some(row_style) = style_for(idx, line) {
-            for _ in 0..rows_for_line {
-                if visual_row >= inner.height as usize {
-                    break;
-                }
-                let row_rect = Rect {
-                    x: inner.x,
-                    y: inner.y + visual_row as u16,
-                    width: inner.width,
-                    height: 1,
-                };
-                frame.buffer_mut().set_style(row_rect, row_style);
-                visual_row += 1;
-            }
-        } else {
-            visual_row += rows_for_line;
-        }
-    }
-}
-
-/// Apply horizontal scroll to a line while preserving the first span (cursor indicator)
-fn apply_horizontal_scroll(line: Line, scroll_x: usize) -> Line {
-    if scroll_x == 0 || line.spans.is_empty() {
-        return line;
-    }
-
-    let mut spans: Vec<Span> = line.spans.into_iter().collect();
-
-    // Preserve the first span (indicator)
-    let indicator = spans.remove(0);
-
-    // Skip scroll_x characters from the remaining spans
-    let mut chars_to_skip = scroll_x;
-    let mut new_spans = vec![indicator];
-
-    for span in spans {
-        let content = span.content.to_string();
-        let char_count = content.chars().count();
-        if chars_to_skip >= char_count {
-            chars_to_skip -= char_count;
-            // Skip this span entirely
-        } else if chars_to_skip > 0 {
-            // Partially skip this span
-            let new_content: String = content.chars().skip(chars_to_skip).collect();
-            chars_to_skip = 0;
-            new_spans.push(Span::styled(new_content, span.style));
-        } else {
-            // Keep this span as-is
-            new_spans.push(Span::styled(content, span.style));
-        }
-    }
-
-    Line::from(new_spans)
-}
-
-#[cfg(test)]
-mod selector_render_snapshot_tests {
-    //! Render-snapshot tests for the review-target selector. We drive the
-    //! real `render` against ratatui's `TestBackend` and assert on the
-    //! resulting character grid (plus a few style checks for the active
-    //! tab highlight). These tests caught a regression where the inactive
-    //! tab rendered as bare dim text with no bracket / cue, making the
-    //! Pull Requests tab functionally invisible.
-    use super::*;
-    use crate::app::DiffSource;
-    use crate::error::Result as TuicrResult;
-    use crate::error::TuicrError;
-    use crate::forge::selector::PullRequestsTab;
-    use crate::forge::traits::{ForgeRepository, PullRequestSummary};
-    use crate::model::{DiffFile, DiffLine, FileStatus, ReviewSession, SessionDiffSource};
-    use crate::syntax::SyntaxHighlighter;
-    use crate::theme::Theme;
-    use crate::vcs::CommitInfo;
-    use crate::vcs::traits::{VcsBackend, VcsChangeStatus, VcsInfo, VcsType};
-    use chrono::{TimeZone, Utc};
-    use ratatui::Terminal;
-    use ratatui::backend::TestBackend;
-    use ratatui::buffer::Buffer;
-    use ratatui::style::Modifier;
-    use std::path::{Path, PathBuf};
-
-    struct SnapshotVcs {
-        info: VcsInfo,
-    }
-
-    impl VcsBackend for SnapshotVcs {
-        fn info(&self) -> &VcsInfo {
-            &self.info
-        }
-
-        fn get_working_tree_diff(
-            &self,
-            _highlighter: &SyntaxHighlighter,
-        ) -> TuicrResult<Vec<DiffFile>> {
-            Err(TuicrError::NoChanges)
-        }
-
-        fn fetch_context_lines(
-            &self,
-            _file_path: &Path,
-            _file_status: FileStatus,
-            _start_line: u32,
-            _end_line: u32,
-        ) -> TuicrResult<Vec<DiffLine>> {
-            Ok(Vec::new())
-        }
-
-        fn get_change_status(&self) -> TuicrResult<VcsChangeStatus> {
-            Ok(VcsChangeStatus {
-                staged: false,
-                unstaged: false,
-            })
-        }
-    }
-
-    fn commit(i: usize) -> CommitInfo {
-        CommitInfo {
-            id: format!("abc{i}"),
-            short_id: format!("abc{i}"),
-            branch_name: None,
-            summary: format!("commit {i}"),
-            body: None,
-            author: "tester".to_string(),
-            time: Utc.with_ymd_and_hms(2026, 1, 1, 0, 0, 0).unwrap(),
-        }
-    }
-
-    fn make_app(commits: Vec<CommitInfo>) -> App {
-        let vcs_info = VcsInfo {
-            root_path: PathBuf::from("/tmp"),
-            head_commit: "head".to_string(),
-            branch_name: Some("main".to_string()),
-            vcs_type: VcsType::Git,
-        };
-        let session = ReviewSession::new(
-            vcs_info.root_path.clone(),
-            vcs_info.head_commit.clone(),
-            vcs_info.branch_name.clone(),
-            SessionDiffSource::WorkingTree,
-        );
-        App::build(
-            Box::new(SnapshotVcs {
-                info: vcs_info.clone(),
-            }),
-            vcs_info,
-            Theme::dark(),
-            None,
-            false,
-            Vec::new(),
-            session,
-            DiffSource::WorkingTree,
-            InputMode::CommitSelect,
-            commits,
-            None,
-        )
-        .expect("build app")
-    }
-
-    fn repo() -> ForgeRepository {
-        ForgeRepository::github("github.com", "agavra", "tuicr")
-    }
-
-    fn pr(number: u64, title: &str, author: &str) -> PullRequestSummary {
-        PullRequestSummary {
-            repository: repo(),
-            number,
-            title: title.to_string(),
-            author: Some(author.to_string()),
-            head_ref_name: format!("feat/{number}"),
-            base_ref_name: "main".to_string(),
-            updated_at: Some(Utc.with_ymd_and_hms(2026, 5, 1, 0, 0, 0).unwrap()),
-            url: format!("https://github.com/agavra/tuicr/pull/{number}"),
-            state: "OPEN".to_string(),
-            is_draft: false,
-        }
-    }
-
-    fn draw(app: &mut App) -> Buffer {
-        let backend = TestBackend::new(120, 24);
-        let mut terminal = Terminal::new(backend).unwrap();
-        terminal
-            .draw(|frame| render(frame, app))
-            .expect("draw frame");
-        terminal.backend().buffer().clone()
-    }
-
-    fn row_text(buffer: &Buffer, y: u16) -> String {
-        (0..buffer.area.width)
-            .map(|x| buffer[(x, y)].symbol().to_string())
-            .collect()
-    }
-
-    /// True when at least one cell in [x_start, x_end) on row `y` carries the
-    /// REVERSED modifier — our visual marker for the active tab.
-    fn any_reversed_in_range(buffer: &Buffer, y: u16, x_start: u16, x_end: u16) -> bool {
-        (x_start..x_end.min(buffer.area.width)).any(|x| {
-            buffer[(x, y)]
-                .style()
-                .add_modifier
-                .contains(Modifier::REVERSED)
-        })
-    }
-
-    /// Locate the inclusive x-range of a substring on row `y`. Panics if
-    /// the substring is not present, with a helpful dump.
-    fn locate(buffer: &Buffer, y: u16, needle: &str) -> (u16, u16) {
-        let line = row_text(buffer, y);
-        let byte_idx = line
-            .find(needle)
-            .unwrap_or_else(|| panic!("expected to find {needle:?} on row {y}, got {line:?}"));
-        // Symbols are ASCII in our render so byte index == column index.
-        let start = byte_idx as u16;
-        let end = start + needle.len() as u16;
-        (start, end)
-    }
-
-    #[test]
-    fn should_render_both_tabs_with_brackets_when_local_active_and_no_forge() {
-        // given — plain app with no GitHub remote
-        let mut app = make_app(vec![commit(0), commit(1)]);
-        // when
-        let buffer = draw(&mut app);
-        // then — tab strip has both bracketed labels
-        let strip = row_text(&buffer, 1);
-        assert!(
-            strip.contains("[ Local ]") && strip.contains("[ Pull Requests ]"),
-            "tab strip missing brackets: {strip:?}"
-        );
-        // and — the active "Local" label is REVERSED, inactive PR label is not
-        let (lo, hi) = locate(&buffer, 1, "Local");
-        assert!(
-            any_reversed_in_range(&buffer, 1, lo, hi),
-            "active Local label should be REVERSED"
-        );
-        let (lo, hi) = locate(&buffer, 1, "Pull Requests");
-        assert!(
-            !any_reversed_in_range(&buffer, 1, lo, hi),
-            "inactive Pull Requests label should NOT be REVERSED"
-        );
-    }
-
-    #[test]
-    fn should_show_disabled_banner_when_pr_tab_active_without_forge() {
-        // given
-        let mut app = make_app(vec![commit(0)]);
-        app.target_tab = crate::app::TargetTab::PullRequests;
-        // when
-        let buffer = draw(&mut app);
-        // then — banner spells out the reason
-        let body = (3..buffer.area.height)
-            .map(|y| row_text(&buffer, y))
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(
-            body.contains("No GitHub remote on this repo"),
-            "expected disabled banner, got body:\n{body}"
-        );
-    }
-
-    #[test]
-    fn should_highlight_pr_tab_label_when_pr_tab_active() {
-        // given — forge repo configured, PR tab active in Idle state
-        let mut app = make_app(vec![commit(0)]);
-        app.forge_repository = Some(repo());
-        app.pr_tab = PullRequestsTab::new(Some(repo()));
-        app.target_tab = crate::app::TargetTab::PullRequests;
-        // when
-        let buffer = draw(&mut app);
-        // then — Pull Requests is the highlighted (REVERSED) label
-        let (lo, hi) = locate(&buffer, 1, "Pull Requests");
-        assert!(
-            any_reversed_in_range(&buffer, 1, lo, hi),
-            "active Pull Requests label should be REVERSED"
-        );
-        let (lo, hi) = locate(&buffer, 1, "Local");
-        assert!(
-            !any_reversed_in_range(&buffer, 1, lo, hi),
-            "inactive Local label should NOT be REVERSED"
-        );
-    }
-
-    #[test]
-    fn should_render_loaded_pr_rows_with_number_title_author() {
-        // given — three loaded PRs, second with no author
-        let mut app = make_app(vec![commit(0)]);
-        app.forge_repository = Some(repo());
-        let mut tab = PullRequestsTab::new(Some(repo()));
-        tab.start_initial_load();
-        tab.apply_initial_load(Ok((
-            vec![
-                pr(148, "Add forge-backed PR review", "alice"),
-                pr(125, "Support fetching/pushing reviews", "ypares"),
-            ],
-            false,
-        )));
-        app.pr_tab = tab;
-        app.target_tab = crate::app::TargetTab::PullRequests;
-        // when
-        let buffer = draw(&mut app);
-        // then — both rows are present in the rendered body
-        let body = (3..buffer.area.height)
-            .map(|y| row_text(&buffer, y))
-            .collect::<Vec<_>>()
-            .join("\n");
-        for needle in [
-            "#148",
-            "Add forge-backed PR review",
-            "alice",
-            "#125",
-            "Support fetching/pushing reviews",
-            "ypares",
-        ] {
-            assert!(body.contains(needle), "missing {needle:?} in:\n{body}");
-        }
-    }
-
-    #[test]
-    fn should_show_load_more_row_when_has_more_and_no_filter() {
-        // given
-        let mut app = make_app(vec![commit(0)]);
-        app.forge_repository = Some(repo());
-        let mut tab = PullRequestsTab::new(Some(repo()));
-        tab.start_initial_load();
-        tab.apply_initial_load(Ok((vec![pr(1, "alpha", "a")], true)));
-        app.pr_tab = tab;
-        app.target_tab = crate::app::TargetTab::PullRequests;
-        // when
-        let buffer = draw(&mut app);
-        // then
-        let body = (3..buffer.area.height)
-            .map(|y| row_text(&buffer, y))
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(
-            body.contains("load more pull requests"),
-            "expected load-more row, body:\n{body}"
-        );
-    }
-
-    #[test]
-    fn should_hide_load_more_row_when_filter_active() {
-        // given — has_more is true but a filter is set
-        let mut app = make_app(vec![commit(0)]);
-        app.forge_repository = Some(repo());
-        let mut tab = PullRequestsTab::new(Some(repo()));
-        tab.start_initial_load();
-        tab.apply_initial_load(Ok((vec![pr(1, "alpha", "a")], true)));
-        tab.set_filter("alpha".to_string());
-        app.pr_tab = tab;
-        app.target_tab = crate::app::TargetTab::PullRequests;
-        // when
-        let buffer = draw(&mut app);
-        // then
-        let body = (3..buffer.area.height)
-            .map(|y| row_text(&buffer, y))
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(
-            !body.contains("load more pull requests"),
-            "load-more row should be hidden while filter is active:\n{body}"
-        );
-    }
-
-    #[test]
-    fn should_render_filter_draft_banner_when_editing_filter() {
-        // given
-        let mut app = make_app(vec![commit(0)]);
-        app.forge_repository = Some(repo());
-        let mut tab = PullRequestsTab::new(Some(repo()));
-        tab.start_initial_load();
-        tab.apply_initial_load(Ok((vec![pr(1, "alpha", "a")], false)));
-        app.pr_tab = tab;
-        app.target_tab = crate::app::TargetTab::PullRequests;
-        app.pr_filter_draft = Some("alp".to_string());
-        // when
-        let buffer = draw(&mut app);
-        // then — the banner shows the in-progress filter expression
-        let body = (3..buffer.area.height)
-            .map(|y| row_text(&buffer, y))
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(
-            body.contains("/alp") && body.contains("filter"),
-            "expected filter draft banner, got:\n{body}"
-        );
-    }
-
-    #[test]
-    fn should_render_error_banner_when_pr_load_failed() {
-        // given
-        let mut app = make_app(vec![commit(0)]);
-        app.forge_repository = Some(repo());
-        let mut tab = PullRequestsTab::new(Some(repo()));
-        tab.start_initial_load();
-        tab.apply_initial_load(Err("network down".to_string()));
-        app.pr_tab = tab;
-        app.target_tab = crate::app::TargetTab::PullRequests;
-        // when
-        let buffer = draw(&mut app);
-        // then
-        let body = (3..buffer.area.height)
-            .map(|y| row_text(&buffer, y))
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(
-            body.contains("error") && body.contains("network down"),
-            "expected error banner, got:\n{body}"
-        );
-    }
-
-    #[test]
-    fn should_render_loading_banner_when_pr_load_in_flight() {
-        // given
-        let mut app = make_app(vec![commit(0)]);
-        app.forge_repository = Some(repo());
-        let mut tab = PullRequestsTab::new(Some(repo()));
-        tab.start_initial_load();
-        // (stays in Loading until apply_initial_load)
-        app.pr_tab = tab;
-        app.target_tab = crate::app::TargetTab::PullRequests;
-        // when
-        let buffer = draw(&mut app);
-        // then
-        let body = (3..buffer.area.height)
-            .map(|y| row_text(&buffer, y))
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(
-            body.contains("Fetching pull requests"),
-            "expected loading banner, got:\n{body}"
-        );
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn should_return_string_unchanged_when_within_max_len() {
-        // given
-        let s = "hello";
-        // when
-        let result = truncate_str(s, 10);
-        // then
-        assert_eq!(result, "hello");
-    }
-
-    #[test]
-    fn should_truncate_ascii_string_with_ellipsis() {
-        // given
-        let s = "hello world this is long";
-        // when
-        let result = truncate_str(s, 10);
-        // then
-        assert_eq!(result, "hello w...");
-    }
-
-    #[test]
-    fn should_truncate_without_panicking_on_multibyte_chars() {
-        // given - the exact string from the bug report
-        let s = "Resolve \"SD : Envoi en validation manuelle après 3 rejet de la fiche employé\"";
-        // when
-        let result = truncate_str(s, 47);
-        // then - should not panic and should end with "..."
-        assert!(result.ends_with("..."));
-        assert!(result.len() <= 47);
-    }
-
-    #[test]
-    fn should_handle_string_of_only_multibyte_chars() {
-        // given
-        let s = "ééééééééé";
-        // when
-        let result = truncate_str(s, 5);
-        // then
-        assert!(result.ends_with("..."));
-        assert!(result.is_char_boundary(result.len()));
-    }
-
-    #[test]
-    fn should_not_scroll_when_comment_box_already_visible() {
-        // given: box at lines 5-7, viewport shows lines 0-9
-        let mut scroll = 0;
-        // when
-        scroll_comment_input_into_view(&mut scroll, Some((5, 7)), Some(6), 10, 100);
-        // then
-        assert_eq!(scroll, 0);
-    }
-
-    #[test]
-    fn should_scroll_down_when_comment_box_below_viewport() {
-        // given: box at lines 20-22, viewport shows lines 0-9
-        let mut scroll = 0;
-        // when
-        scroll_comment_input_into_view(&mut scroll, Some((20, 22)), Some(21), 10, 100);
-        // then: scroll so box_end (22) is the last visible line => scroll = 22 - 10 + 1 = 13
-        assert_eq!(scroll, 13);
-    }
-
-    #[test]
-    fn should_scroll_up_when_comment_box_above_viewport() {
-        // given: box at lines 5-7, viewport shows lines 20-29
-        let mut scroll = 20;
-        // when
-        scroll_comment_input_into_view(&mut scroll, Some((5, 7)), Some(6), 10, 100);
-        // then: scroll so box_start (5) is the first visible line
-        assert_eq!(scroll, 5);
-    }
-
-    #[test]
-    fn should_scroll_to_cursor_when_box_taller_than_viewport() {
-        // given: box spans 20 lines, viewport only 10 lines
-        let mut scroll = 0;
-        // when
-        scroll_comment_input_into_view(&mut scroll, Some((30, 49)), Some(45), 10, 100);
-        // then: scroll so cursor (45) is the last visible line => scroll = 45 - 10 + 1 = 36
-        assert_eq!(scroll, 36);
-    }
-
-    #[test]
-    fn should_not_scroll_past_end_of_content() {
-        // given: scroll already past max (e.g., content shrank)
-        let mut scroll = 200;
-        // when
-        scroll_comment_input_into_view(&mut scroll, Some((95, 97)), Some(96), 10, 100);
-        // then: clamped to max_scroll = 100 - 10 = 90
-        assert_eq!(scroll, 90);
-    }
-
-    #[test]
-    fn should_not_scroll_when_no_comment_box() {
-        // given
-        let mut scroll = 42;
-        // when
-        scroll_comment_input_into_view(&mut scroll, None, None, 10, 100);
-        // then
-        assert_eq!(scroll, 42);
-    }
-
-    #[test]
-    fn should_handle_box_partially_below_viewport() {
-        // given: viewport shows 0-9, box starts at 8 and ends at 10 (footer off-screen)
-        let mut scroll = 0;
-        // when
-        scroll_comment_input_into_view(&mut scroll, Some((8, 10)), Some(9), 10, 100);
-        // then: scroll so box_end (10) is visible => scroll = 10 - 10 + 1 = 1
-        assert_eq!(scroll, 1);
-    }
-
-    #[test]
-    fn should_pad_highlighted_spans_to_exact_width() {
-        // given - highlighted spans from the syntax highlighter (which strips
-        // the trailing \n that syntect includes). Short content gets padded
-        // by truncate_or_pad_spans; the result must have exactly `width`
-        // characters so the side-by-side separator stays aligned.
-        let highlighter = crate::syntax::SyntaxHighlighter::default();
-        let lines = vec!["let x = 1;".to_string()];
-        let highlighted = highlighter
-            .highlight_file_lines(std::path::Path::new("test.rs"), &lines)
-            .unwrap();
-        let spans = highlighted[0].as_ref().unwrap();
-
-        let width = 80;
-
-        // when
-        let result = truncate_or_pad_spans(spans, width, Style::default());
-
-        // then - total char count must equal the target width so each
-        // side-by-side column is the same size
-        let total_chars: usize = result.iter().map(|s| s.content.chars().count()).sum();
-        assert_eq!(
-            total_chars, width,
-            "padded spans should have exactly {width} chars, got {total_chars}"
-        );
     }
 }
