+++ src/ui/mod.rs
@@ -1,9 +1,14 @@
 pub mod app_layout;
 pub mod comment_panel;
+pub mod diff_side_by_side;
+pub mod diff_unified;
 pub mod diff_view;
 pub mod file_list;
 pub mod help_popup;
+pub mod inline_commit_selector;
+pub mod selector;
 pub mod status_bar;
 pub mod styles;
+pub mod text_utils;
 
 pub use app_layout::render;
